### 2024-10-14 #7352 @editorialbot
[PRE REVIEW]: OPTIMUS: a multidimensional global optimization package 
**Submitting author:** @itsoulos (<a href="http://orcid.org/0000-0002-2343-2733">Ioannis Tsoulos</a>)
**Repository:** https://github.com/itsoulos/GlobalOptimus/
**Branch with paper.md** (empty if default branch):
**Version:** 1.0
