### 2024-10-14 #7352 @danielskatz
👋 @itsoulos - Also, I notice your paper is about 5x as long as a JOSS paper should be. Please see the example paper for this as well, and reduce your paper to be more like the 250-1000 words that JOSS expects. You can possible put a lot of text elsewhere in the repo or docs, and then point to it there. Alternatively, it could be that JOSS is not the right venue for your publication, which would be ok, too, but if so, please tell me so I can mark this submission as withdrawn. You can use the command `@editorialbot check repository` to check a few things, including the length of the paper
