+++ paper/paper.md
@@ -6,511 +6,76 @@
 ## Summary
 
 <p style="text-align: justify;">
-A significant number of applications from many research areas can be considered global optimization problems, such as applications in the area of image processing, medical informatics, economic models, etc. This paper presents a programming tool written in ANSI C++, which researchers can use to formulate the problem to be solved and then make use of the local and global optimization methods provided by this tool to efficiently solve such problems. The main features of the suggested software are: a) Coding of the objective problem in a high level language such as ANSI C++ b) Incorporation of many global optimization techniques to tackle the objective problem c)Parameterization of global optimization methods using user-defined parameters d) Usage of a GUI application to control the optimization strategy. 
+This paper presents a programming tool written in ANSI C++, which researchers can use to formulate the problem to be solved and then make use of the local and global optimization methods provided by this tool to efficiently solve such problems. The main features of the suggested software are: a) Coding of the objective problem in a high level language such as ANSI C++ b) Incorporation of many global optimization techniques to tackle the objective problem c)Parameterization of global optimization methods using user-defined parameters d) Usage of a GUI application to control the optimization strategy. 
 </p>
 <b>Keywords:</b> Global optimization; stochastic methods; Genetic algorithms; Software;
 
-## 1. Introduction 
+##  Software 
 <p style="text-align: justify;">
-The location of the global minimum for a continuous and differentiable function f : S → R, S ⊂ Rn is formulated as 
-x = arg min x∈S f(x) (1)
-<p style="text-align: justify;">
-where the set S is defined as: 
- S = [a1, b1] ⊗ [a2, b2] ⊗ . . . [an, bn]
-</p>
-<p style="text-align: justify;">
-Methods that aim to locate the global minimum finds application in problems from the area of economics, problems that appear very often in the area of physics, chemistry, common problems from medicine, job scheduling problems, water resources planning, network security problems, robotics etc. Also, global optimization methods were used on some symmetry problems  as well as on inverse problems. In the relevant literature there are a number of global optimization techniques, such as Adaptive Random Search methods, Controlled Random Search methods, Simulated Annealing, Genetic algorithms, Ant Colony Optimization, Particle Swarm Optimization  etc. Due to the high importance of the global optimization problem, a variety of hybrid optimization techniques have been proposed to handle the global optimization problem, such as methods that combine Particle Swarm Optimization and Genetic algorithms, combination of genetic algorithms and fuzzy logic classifier, incorporation of genetic algorithm and the K-Means algorithm, combination of Particle Swarm Optimization method with Ant Colony Optimization, methods that combine the Simplex method and Inductive search etc. Also, many hybrid techniques combining local and global optimization have been developed. Just a few recent application examples include an adaptive genetic algorithm for crystal structure prediction, modeling of fusion plasma physics with genetic algorithms, usage of genetic algorithms for astroparticle physics studies, parameter extraction of solar cells using a Particle Swarm Optimization method, a new control approach of a fleet of Unmanned Aerial Vehicles using the method of Particle Swarm Optimization etc. However, in most cases, global optimization methods require a lot of computing resources to implement both in memory and computing time. Because of the large demands that global optimization methods have on computing power, several techniques have been proposed, such as asynchronous methods, parallel approaches of the Multistart optimization method and also some methods that take advantage of modern parallel GPU architectures. 
-</p>
-<p style="text-align: justify;">
[…]
