+++ README.md
@@ -30,6 +30,68 @@ Alternatively, in Windows the user can use the added Setup msi executable and is
 2. Εdit the 'settings' shortcut on the desktop and adjust the optimizer according to the comments
 3. Open the command prompt (cmd) or PowerShell and to optimize, you can write the objective function you desire. For example, you can type 'runfunmin camel' or 'runfunmin potential 10' and so on.
 
+
+
+<p style="text-align: justify;">
+The steps of the installation are shown in Figure 1. The user needs to select only the desired installation directory and the installer copies the necessary files to this directory.
+</p>
+
+(a) First screen of the windows installation wizard.
+![win_1.png](paper/win_1.png)
+
+(b) The user selects the desired installation directory
+![win_2.png](paper/win_2.png)
+
+
+(c) Prompt to install.
+![win_3](paper/win_3.png)
+(d) Copying files.
+![win_4](paper/win_4.png)
+(e) Finalizing installation.
+![win_5](paper/win_5.png)
+**Figure 1.** The steps of windows installation
+
+
+##  The GUI application 
+<p style="text-align: justify;">
+The software also provides a graphical interface with which the user can easily select an objective function, optimization method and modify them by changing a number of available parameters. The software is called Xoptimus and the user can compile this program using the commands: 
+</p>
+
+1. qmake Xoptimus.pro 
+2. make 
+<p style="text-align: justify;">
+Also, in Windows environments the Xoptimus.msi package will install this program. After invoking the program the first screen that is appeared is shown in Figure 6.
+</p>
+
+![xoptimus_start](paper/xoptimus_start.png)
+
+
+**Figure 6.** The start screen of Xoptimus program.
+<p style="text-align: justify;">
+The user should select first the LOAD menu in order to define the objective problem and the desired global optimization method that should be used. The screen of Figure 7 is shown when the user select the option PROBLEM from the LOAD menu.
+</p>
+
+![xoptimus_problem_selection](paper/xoptimus_problem_selection.png)
+
+
+**Figure 7.** Selecting an optimization problem.
+<p style="text-align: justify;">
+The user can select from a variety of offered objective functions from the dropdown menu. For each selected function the bounds are shown in the screen. Having selected the objective problem, the user should also selects the desired global optimization method from the option METHOD of the LOAD menu. This choice shows the dialog of Figure 8 in the screen. 
+</p>
+
+![xoptimus_method_selection](paper/xoptimus_method_selection.png)
+
+
+**Figure 8.** Selecting an optimization method.
+<p style="text-align: justify;">
+From this dialog, the user can select the global optimization method from a dropdown menu and the user can also modify some of the parameters of the method, such as the number of chromosomes in the Genetic algorithm. Having selected the function and the desired global optimization method, the user can start the optimization process by invoking the option RUN from the menu EXECUTE. The selected global optimization method initiates and an example of the execution is shown in Figure 9.
+</p>
+
+![xoptimus_rastrigin_genetic](paper/xoptimus_rastrigin_genetic.png)
+
+
+**Figure 9.** Executing the selected optimization method.
+
 ## Available Global Optimization methods
 The implemented Global Optimization methods are:
 1. The Differential Evolution denoted as DifferentialEvolution.
@@ -38,8 +100,11 @@ The implemented Global Optimization methods are:
 4. An improved version of the Particle Swarm Optimizer, denoted as iPSO
 5. A simple Multistart method, which initiates many local optimization procedures to discover the global optimum of a function.
 6. The NeuralMinimizer method, which constructs an estimated function of the objective problem using Neural Networks.
[…]
