### 2024-10-16 #7352 @itsoulos
Dear editor,
We have again upload the paper in the relevant github repository and we have reduced the size to the required one.
Also, in order to compile the GlobalOptimus the used should follow the instructions provided in the README.md
In most unix systems the user should invoke the following commands:
1. qmake GlobalOptimus.pro
2. make
Also, we have included msi files for Windows users, where no compilation is required

Best regards,
Ioannis G. Tsoulos
Professor
