### 2024-10-16 #7352 @danielskatz
👋 @itsoulos - you now need to keep working on the paper until it compiles.  The word count is good, however.
