### 2024-10-16 #7352 @itsoulos
Dear editor,
in order to compile the package you need the QT library as we mention in the git repository. After that, the user should invoke two commands
1. qmake GlobalOptimus.pro
2. make
