### 2024-10-16 5e92aca
small fixes in paper
