### 2024-10-16 22d781c
small fixes in paper
