### 2024-10-16 e235b77
small fixes in paper
