### 2024-10-16 d93e0d2
small fixes in paper
