+++ paper/paper.bib
@@ -1,115 +1,107 @@
-{\rtf1\ansi\ansicpg1253\cocoartf2639
-\cocoatextscaling0\cocoaplatform0{\fonttbl\f0\fmodern\fcharset0 Courier;}
-{\colortbl;\red255\green255\blue255;\red0\green0\blue0;}
-{\*\expandedcolortbl;;\cssrgb\c0\c0\c0;}
-\paperw11900\paperh16840\margl1440\margr1440\vieww28600\viewh18000\viewkind0
-\deftab720
-\pard\pardeftab720\partightenfactor0
+@article{tsoulos2016pdoublepop,
+  title={PDoublePop: An implementation of parallel genetic algorithm for function optimization},
+  author={Tsoulos, Ioannis G and Tzallas, Alexandros and Tsalikakis, Dimitris},
+  journal={Computer Physics Communications},
+  volume={209},
+  pages={183--189},
+  year={2016},
+  publisher={Elsevier}
+}
 
-\f0\fs26 \cf0 \expnd0\expndtw0\kerning0
-@article\{tsoulos2016pdoublepop,\
-  title=\{PDoublePop: An implementation of parallel genetic algorithm for function optimization\},\
-  author=\{Tsoulos, Ioannis G and Tzallas, Alexandros and Tsalikakis, Dimitris\},\
-  journal=\{Computer Physics Communications\},\
-  volume=\{209\},\
-  pages=\{183--189\},\
-  year=\{2016\},\
-  publisher=\{Elsevier\}\
-\}\
-\
-\pard\pardeftab720\partightenfactor0
-\cf0 @article\{charilogis2023parallel,\
-  title=\{A parallel implementation of the differential evolution method\},\
-  author=\{Charilogis, Vasileios and Tsoulos, Ioannis G\},\
-  journal=\{Analytics\},\
-  volume=\{2\},\
-  number=\{1\},\
-  pages=\{17--30\},\
-  year=\{2023\},\
-  publisher=\{MDPI\}\
-\}\
-\
-\
-@article\{tsoulos2008modifications,\
-  title=\{Modifications of real code genetic algorithm for global optimization\},\
-  author=\{Tsoulos, Ioannis G\},\
-  journal=\{Applied Mathematics and Computation\},\
-  volume=\{203\},\
-  number=\{2\},\
-  pages=\{598--607\},\
-  year=\{2008\},\
-  publisher=\{Elsevier\}\
-\}\
-\
-\
-@article\{charilogis2022toward,\
-  title=\{Toward an ideal particle swarm optimizer for multidimensional functions\},\
-  author=\{Charilogis, Vasileios and Tsoulos, Ioannis G\},\
-  journal=\{Information\},\
-  volume=\{13\},\
-  number=\{5\},\
-  pages=\{217\},\
-  year=\{2022\},\
-  publisher=\{MDPI\}\
-\}\
-\
-\
-\
-@article\{tsoulos2023neuralminimizer,\
-  title=\{NeuralMinimizer: A Novel Method for Global Optimization\},\
-  author=\{Tsoulos, Ioannis G and Tzallas, Alexandros and Karvounis, Evangelos and Tsalikakis, Dimitrios\},\
-  journal=\{Information\},\
-  volume=\{14\},\
-  number=\{2\},\
-  pages=\{66\},\
-  year=\{2023\},\
-  publisher=\{MDPI\}\
-\}\
-\
-@article\{charilogis2023improved,\
-  title=\{An improved parallel particle swarm optimization\},\
-  author=\{Charilogis, Vasileios and Tsoulos, Ioannis G and Tzallas, Alexandros\},\
[…]
