### 2024-10-16 cdc01cc
small fixes in paper
