+++ paper/paper.bib
@@ -8,7 +8,7 @@
   publisher={Elsevier}
 }
 
-@article{pant2020differential,
+@article{pant2020differential_2020,
   title={Differential Evolution: A review of more than two decades of research},
   author={Pant, Millie and Zaheer, Hira and Garcia-Hernandez, Laura and Abraham, Ajith and others},
   journal={Engineering Applications of Artificial Intelligence},
@@ -18,7 +18,7 @@
   publisher={Elsevier}
 }
 
-@article{charilogis2023parallel,
+@article{charilogis2023parallel_2023,
   title={A parallel implementation of the differential evolution method},
   author={Charilogis, Vasileios and Tsoulos, Ioannis G},
   journal={Analytics},
