### 2024-10-16 a7a1c27
small fixes in paper
