+++ paper/paper.bib
@@ -29,7 +29,7 @@
   publisher={MDPI}
 }
 
-@article{tsoulos2008modifications,
+@article{tsoulos2008modifications_2008,
   title={Modifications of real code genetic algorithm for global optimization},
   author={Tsoulos, Ioannis G},
   journal={Applied Mathematics and Computation},
@@ -40,7 +40,7 @@
   publisher={Elsevier}
 }
 
-@article{charilogis2022toward,
+@article{charilogis2022toward_2022,
   title={Toward an ideal particle swarm optimizer for multidimensional functions},
   author={Charilogis, Vasileios and Tsoulos, Ioannis G},
   journal={Information},
@@ -51,7 +51,7 @@
   publisher={MDPI}
 }
 
-@article{tsoulos2023neuralminimizer,
+@article{tsoulos2023neuralminimizer_2023,
   title={NeuralMinimizer: A Novel Method for Global Optimization},
   author={Tsoulos, Ioannis G and Tzallas, Alexandros and Karvounis, Evangelos and Tsalikakis, Dimitrios},
   journal={Information},
