### 2024-10-16 e83c2e8
small fixes in paper
