+++ paper/paper.bib
@@ -62,7 +62,7 @@
   publisher={MDPI}
 }
 
-@article{charilogis2023improved,
+@article{charilogis2023improved_2023,
   title={An improved parallel particle swarm optimization},
   author={Charilogis, Vasileios and Tsoulos, Ioannis G and Tzallas, Alexandros},
   journal={SN Computer Science},
@@ -73,7 +73,18 @@
   publisher={Springer}
 }
 
-@article{kyrou2024eofa,
+@article{kirkpatrick1983optimization_1983,
+  title={Optimization by simulated annealing},
+  author={Kirkpatrick, Scott and Gelatt Jr, C Daniel and Vecchi, Mario P},
+  journal={science},
+  volume={220},
+  number={4598},
+  pages={671--680},
+  year={1983},
+  publisher={American association for the advancement of science}
+}
+
+@article{kyrou2024eofa_2024,
   title={EOFA: An Extended Version of the Optimal Foraging Algorithm for Global Optimization Problems},
   author={Kyrou, Glykeria and Charilogis, Vasileios and Tsoulos, Ioannis G},
   journal={Computation},
@@ -84,7 +95,7 @@
   publisher={MDPI}
 }
 
-@article{kyrou2024improving,
+@article{kyrou2024improving_2024,
   title={Improving the Giant-Armadillo Optimization Method},
   author={Kyrou, Glykeria and Charilogis, Vasileios and Tsoulos, Ioannis G},
   journal={Analytics},
