### 2024-10-16 0af10ba
small fixes in paper
