+++ paper/paper.bib
@@ -95,6 +95,89 @@
   publisher={MDPI}
 }
 
+@article{powell1989tolerant_1989,
+  title={A tolerant algorithm for linearly constrained optimization calculations},
+  author={Powell, MJD},
+  journal={Mathematical Programming},
+  volume={45},
+  pages={547--566},
+  year={1989},
+  publisher={Springer}
+}
+
+@article{amari1993backpropagation_1993,
+  title={Backpropagation and stochastic gradient descent method},
+  author={Amari, Shun-ichi},
+  journal={Neurocomputing},
+  volume={5},
+  number={4-5},
+  pages={185--196},
+  year={1993},
+  publisher={Elsevier}
+}
+
+@article{kingma2014adam_2014,
+  title={Adam: A method for stochastic optimization},
+  author={Kingma, Diederik P},
+  journal={arXiv preprint arXiv:1412.6980},
+  year={2014}
+}
+
+
+@article{olsson1975nelder_1975,
+  title={The Nelder-Mead simplex procedure for function minimization},
+  author={Olsson, Donald M and Nelson, Lloyd S},
+  journal={Technometrics},
+  volume={17},
+  number={1},
+  pages={45--51},
+  year={1975},
+  publisher={Taylor \& Francis}
+}
+
+@article{liu1989limited_1989,
+  title={On the limited memory BFGS method for large scale optimization},
+  author={Liu, Dong C and Nocedal, Jorge},
+  journal={Mathematical programming},
+  volume={45},
+  number={1},
+  pages={503--528},
+  year={1989},
+  publisher={Springer}
+}
+
+@article{duan1992effective_1992,
+  title={Effective and efficient global optimization for conceptual rainfall-runoff models},
+  author={Duan, Qingyun and Sorooshian, Soroosh and Gupta, Vijai},
+  journal={Water resources research},
+  volume={28},
+  number={4},
+  pages={1015--1031},
+  year={1992},
+  publisher={Wiley Online Library}
+}
+
+@article{li2021improved_2021,
+  title={An improved gray wolf optimization algorithm to solve engineering problems},
+  author={Li, Yu and Lin, Xiaoxiao and Liu, Jingsen},
+  journal={Sustainability},
+  volume={13},
+  number={6},
+  pages={3208},
+  year={2021},
+  publisher={MDPI}
+}
+
+@incollection{marti2010advanced_2010,
+  title={Advanced multi-start methods},
+  author={Mart{\'\i}, Rafael and Moreno-Vega, J Marcos and Duarte, Abraham},
+  booktitle={Handbook of metaheuristics},
+  pages={265--281},
+  year={2010},
+  publisher={Springer}
+}
+
+
 @article{kyrou2024improving_2024,
   title={Improving the Giant-Armadillo Optimization Method},
   author={Kyrou, Glykeria and Charilogis, Vasileios and Tsoulos, Ioannis G},
