+++ paper/paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 ---
 
 ## Summary
-The location of global minimum founds application in a variety of problems derived from physics, chemistry, medicine etc. Also, the global optimization process can be used in complex problems, such as the training of neural networks. Methods used in the global optimization process can be divided into stochastic and deterministic and this article proposes a software that incorporates a variety of stochastic methods.
+The location of global minimum founds application in a variety of problems derived from physics [@duan1992effective_1992], chemistry, medicine etc. Also, the global optimization process can be used in complex problems, such as the training of neural networks. Methods used in the global optimization process can be divided into stochastic and deterministic and this article proposes a software that incorporates a variety of stochastic methods.
 
 
 # Statement of need
@@ -42,7 +42,7 @@ This paper presents the `GlobalOptimus` software tool,  written in ANSI C++, whi
 
 ##  Software 
 <p style="text-align: justify;">
-The suggested software is entirely coded in ANSI C++, using  programming library available from https://qt.io. The researcher should code the objective function and a number of other mandatory functions in the C++ programming language. Also, the researcher should provide the dimension of the objective function as well as the bound of the function. Subsequently, the user can select a global optimization method to apply to the problem from a wide range of available methods.</p>
+The software uses the QT library, available from  https://qt.io. The researcher should code the objective function and a number of other mandatory functions in the C++ programming language. Also, the researcher should provide the dimension of the objective function as well as the bound of the function. Subsequently, the user can select a global optimization method to apply to the problem from a wide range of available methods.</p>
 
 ### Implemented global optimization methods 
 <p style="text-align: justify?">
@@ -55,13 +55,13 @@ METHOD |  DESCRIPTION |
 | **Parallel Differential Evolution**| A parallel version of the Differential Evolution method [@charilogis2023parallel_2023].|
 | **Double precision genetic algorithm**  | An Improved version of the Genetic Algorithm [@tsoulos2008modifications_2008].|
 | **Improved Particle Swarm Optimization**  |Suggested as an improvement of the PSO method [@charilogis2022toward_2022].|
-| **Multistart**  |Initiates local searches from different start points.|
+| **Multistart**  |Initiates local searches from different start points [@marti2010advanced_2010].|
 | **NeuralMinimizer**| Utilizes Radial Basis Functions (RBF) networks to increase the speed of the Multistart method [@tsoulos2023neuralminimizer_2023].|
 | **Parallel Particle Swarm optimizer**  | A novell parallel PSO variant, published recently [@charilogis2023improved_2023].|
 | **Simulated annealing optimizer**  | included in the software under the name Simman [@kirkpatrick1983optimization_1983]. |
 | **The optimal foraging algorithm (OFA)**  | motivated by animal behavioral ecology included the software named Ofa [@kyrou2024eofa_2024].|
 | **Bio-inspired metaheuristic algorithm Giant Armadillo Optimization (GAO)** |  This process mimics the natural behavior of the giant armadillo [@kyrou2024improving_2024]. |
-| **The Gray Wolf Optimizer (GWO)** |  mimics hunting mechanism of gray wolves in nature  included in the software named Gwo. |
+| **The Gray Wolf Optimizer (GWO)** |  This process mimics hunting mechanism of gray wolves in nature [@li2021improved_2021].|
 
 ###  Implemented local optimization methods 
 <p style="text-align: justify;">
@@ -70,11 +70,11 @@ The parameter used to determine the used local optimization procedure is the −
 
 METHOD |  DESCRIPTION |
 |----------|----------|
[…]
