### 2024-10-16 83ff72f
small fixes in paper
