+++ paper/paper.md
@@ -32,7 +32,9 @@ bibliography: paper.bib
 ---
 
 ## Summary
+<p style="text-align: justify;">
 The location of global minimum founds application in a variety of problems derived from physics [@duan1992effective_1992], chemistry, medicine etc. Also, the global optimization process can be used in complex problems, such as the training of neural networks. Methods used in the global optimization process can be divided into stochastic and deterministic and this article proposes a software that incorporates a variety of stochastic methods.
+</p>
 
 
 # Statement of need
@@ -89,10 +91,10 @@ Method addParam()  |  DESCRIPTION |
 |    **done()**    |   This function will be called when the optimization method terminates.|                                       
 |   **~UserMethod()**     |    This is the destructor of the optimization method.  |                                            
 
-A flowchart of any used optimization method is outlined in Figure \autoref{userMethod}.
+A flowchart of any used optimization process is outlined in Figure \autoref{FIG:optimization}.
 
 
-![usermethod] \label{userMethod}(usermethod.png)
+![The diagram of the optimization process \label{FIG:optimization}](userMethod.png)
 
 
 
