### 2024-10-16 843d48c
small fixes in paper
