### 2024-10-16 528462d
small fixes in paper
