### 2024-10-16 9f53f13
small fixes in paper
