+++ paper/paper.bib
@@ -145,6 +145,36 @@
   year={1989},
   publisher={Springer}
 }
+@article{lee2007large_2007,
+  title={Large-scale optimization-based classification models in medicine and biology},
+  author={Lee, Eva K},
+  journal={Annals of biomedical engineering},
+  volume={35},
+  pages={1095--1109},
+  year={2007},
+  publisher={Springer}
+}
+@article{chiroma2017neural_2017,
+  title={Neural networks optimization through genetic algorithm searches: a review},
+  author={Chiroma, Haruna and Noor, Ahmad Shukri Mohd and Abdulkareem, Sameem and Abubakar, Adamu I and Hermawan, Arief and Qin, Hongwu and Hamza, Mukhtar Fatihu and Herawan, Tutut},
+  journal={Appl. Math. Inf. Sci},
+  volume={11},
+  number={6},
+  pages={1543--1564},
+  year={2017}
+}
+
+
+@article{liwo1999protein_1999,
+  title={Protein structure prediction by global optimization of a potential energy function},
+  author={Liwo, Adam and Lee, Jooyoung and Ripoll, Daniel R and Pillardy, Jaroslaw and Scheraga, Harold A},
+  journal={Proceedings of the National Academy of Sciences},
+  volume={96},
+  number={10},
+  pages={5482--5485},
+  year={1999},
+  publisher={National Acad Sciences}
+}
 
 @article{duan1992effective_1992,
   title={Effective and efficient global optimization for conceptual rainfall-runoff models},
@@ -188,6 +218,17 @@
   year={2024},
   publisher={MDPI}
 }
+@article{sergeyev2018efficiency_2018,
+  title={On the efficiency of nature-inspired metaheuristics in expensive global optimization with limited budget},
+  author={Sergeyev, Ya D and Kvasov, DE and Mukhametzhanov, MS},
+  journal={Scientific reports},
+  volume={8},
+  number={1},
+  pages={453},
+  year={2018},
+  publisher={Nature Publishing Group UK London}
+}
+
 
 @article{tsoulos2008genmin,
   title={GenMin: An enhanced genetic algorithm for global optimization},
