+++ paper/paper.bib
@@ -237,6 +237,6 @@
   volume={178},
   number={11},
   pages={843--851},
-  year={2008},\
+  year={2008},
   publisher={Elsevier}
 }
