### 2024-10-16 #7352 @itsoulos
Dear editor,
I have managed to create the proof!
Is this ok now?
