### 2024-10-16 #7352 @danielskatz
@itsoulos - yes, though I will also check the references now to make sure those with DOIs do have the DOIs listed
