### 2024-10-16 #7352 @danielskatz
@itsoulos - it looks like a lot of your references might need DOIs. Please add them to the .bib file, e.g. `doi={10.1016/j.cpc.2016.09.006},` and then re-run the check references command
