+++ paper/paper.bib
@@ -1,13 +1,3 @@
-@article{tsoulos2016pdoublepop,
-  title={PDoublePop: An implementation of parallel genetic algorithm for function optimization},
-  author={Tsoulos, Ioannis G and Tzallas, Alexandros and Tsalikakis, Dimitris},
-  journal={Computer Physics Communications},
-  volume={209},
-  pages={183--189},
-  year={2016},
-  publisher={Elsevier}
-}
-
 @article{pant2020differential_2020,
   title={Differential Evolution: A review of more than two decades of research},
   author={Pant, Millie and Zaheer, Hira and Garcia-Hernandez, Laura and Abraham, Ajith and others},
@@ -15,6 +5,7 @@
   volume={90},
   pages={103479},
   year={2020},
+  doi={10.1016/j.engappai.2020.103479},
   publisher={Elsevier}
 }
 
@@ -26,6 +17,7 @@
   number={1},
   pages={17--30},
   year={2023},
+  doi={10.3390/analytics2010002},
   publisher={MDPI}
 }
 
@@ -37,6 +29,7 @@
   number={2},
   pages={598--607},
   year={2008},
+  doi={10.1016/j.amc.2008.05.005},
   publisher={Elsevier}
 }
 
@@ -48,6 +41,7 @@
   number={5},
   pages={217},
   year={2022},
+  doi={10.3390/info13050217},
   publisher={MDPI}
 }
 
@@ -59,6 +53,7 @@
   number={2},
   pages={66},
   year={2023},
+  doi={10.3390/info14020066 },
   publisher={MDPI}
 }
 
@@ -70,6 +65,7 @@
   number={6},
   pages={766},
   year={2023},
+  doi={10.1007/s42979-023-02227-9},
   publisher={Springer}
 }
 
@@ -81,6 +77,7 @@
   number={4598},
   pages={671--680},
   year={1983},
+  doi={10.1126/science.220.4598.671},
   publisher={American association for the advancement of science}
 }
 
@@ -92,6 +89,7 @@
   number={8},
   pages={158},
   year={2024},
+  doi={10.3390/computation12080158 },
   publisher={MDPI}
 }
 
@@ -102,6 +100,7 @@
   volume={45},
   pages={547--566},
   year={1989},
+  doi={10.1007/bf01589118},
   publisher={Springer}
 }
 
@@ -113,6 +112,7 @@
   number={4-5},
   pages={185--196},
   year={1993},
+  doi={10.1016/0925-2312(93)90006-o},
   publisher={Elsevier}
 }
 
@@ -120,6 +120,7 @@
   title={Adam: A method for stochastic optimization},
   author={Kingma, Diederik P},
   journal={arXiv preprint arXiv:1412.6980},
+  doi={10.48550/arXiv.1412.6980},
   year={2014}
 }
 
@@ -132,6 +133,7 @@
   number={1},
   pages={45--51},
   year={1975},
+  doi={10.2307/1267998},
   publisher={Taylor \& Francis}
 }
 
@@ -143,6 +145,7 @@
   number={1},
   pages={503--528},
   year={1989},
+  doi={10.1007/bf01589116},
   publisher={Springer}
 }
 @article{lee2007large_2007,
@@ -152,6 +155,7 @@
   volume={35},
   pages={1095--1109},
   year={2007},
+  doi={10.1007/s10439-007-9317-7},
   publisher={Springer}
 }
 @article{chiroma2017neural_2017,
@@ -161,6 +165,7 @@
   volume={11},
   number={6},
   pages={1543--1564},
+  doi={10.18576/amis/110602 },
   year={2017}
 }
 
@@ -173,6 +178,7 @@
   number={10},
   pages={5482--5485},
   year={1999},
[…]
