+++ Makefile
@@ -1,9 +1,9 @@
 #############################################################################
-# Makefile for building: Xoptimus
-# Generated by qmake (3.1) (Qt 5.15.8)
-# Project:  Xoptimus.pro
+# Makefile for building: DataFitting
+# Generated by qmake (3.1) (Qt 6.7.2)
+# Project:  DataFitting.pro
 # Template: app
-# Command: /usr/lib/qt5/bin/qmake -o Makefile Xoptimus.pro
+# Command: /usr/bin/qmake -o Makefile DataFitting.pro
 #############################################################################
 
 MAKEFILE      = Makefile
@@ -15,10 +15,10 @@ EQ            = =
 CC            = gcc
 CXX           = g++
 DEFINES       = -DQT_NO_DEBUG -DQT_WIDGETS_LIB -DQT_GUI_LIB -DQT_CORE_LIB
-CFLAGS        = -pipe -O2 -O3 -march=native -fopenmp -unroll-loops -omit-frame-pointer -Winline -unsafe-math-optimizations -mtune=native -pthread -Wall -Wextra -D_REENTRANT -fPIC $(DEFINES)
-CXXFLAGS      = -pipe -O2 -O3 -fopenmp -unroll-loops -fomit-frame-pointer -Winline -unsafe-math-optimizations -mtune=native -pthread -std=gnu++1z -Wall -Wextra -D_REENTRANT -fPIC $(DEFINES)
-INCPATH       = -I. -I/usr/include/x86_64-linux-gnu/qt5 -I/usr/include/x86_64-linux-gnu/qt5/QtWidgets -I/usr/include/x86_64-linux-gnu/qt5/QtGui -I/usr/include/x86_64-linux-gnu/qt5/QtCore -I. -I/usr/lib/x86_64-linux-gnu/qt5/mkspecs/linux-g++
-QMAKE         = /usr/lib/qt5/bin/qmake
+CFLAGS        = -pipe -O3 -fopenmp -unroll-loops -omit-frame-pointer -Winline -unsafe-math-optimizations -mtune=native -pthread -O2 -O3 -march=native -fopenmp -unroll-loops -omit-frame-pointer -Winline -unsafe-math-optimizations -mtune=native -pthread -Wall -Wextra -D_REENTRANT $(DEFINES)
+CXXFLAGS      = -pipe -O3 -fopenmp -unroll-loops -omit-frame-pointer -Winline -unsafe-math-optimizations -mtune=native -pthread -O2 -O3 -fopenmp -unroll-loops -omit-frame-pointer -Winline -unsafe-math-optimizations -mtune=native -pthread -std=gnu++1z -Wall -Wextra -D_REENTRANT $(DEFINES)
+INCPATH       = -I. -I/usr/include/qt6 -I/usr/include/qt6/QtWidgets -I/usr/include/qt6/QtGui -I/usr/include/qt6/QtCore -I. -I/usr/lib64/qt6/mkspecs/linux-g++
+QMAKE         = /usr/bin/qmake
 DEL_FILE      = rm -f
 CHK_DIR_EXISTS= test -d
 MKDIR         = mkdir -p
@@ -28,19 +28,19 @@ COPY_DIR      = cp -f -R
 INSTALL_FILE  = install -m 644 -p
[…]
