### 2024-10-19 3c2ceb6
hybrid doc
