### 2024-10-19 5498c6e
start adding fc model
