+++ MLMODELS/model.h
@@ -2,6 +2,7 @@
 #define MODEL_H
 # include <MLMODELS/dataset.h>
 # include <OPTIMUS/parameterlist.h>
+# include <OPTIMUS/optimizer.h>
 class Model : public ParameterList
 {
 protected:
@@ -9,6 +10,7 @@ protected:
     Dataset *testDataset;
     bool    noRemoveData=false;
     Matrix xall;
+    Optimizer *method;
 public:
     Model();
     void        disableRemoveData();
@@ -18,12 +20,15 @@ public:
     double      getTestError(Dataset *test) ;
     double      getClassTestError(Dataset *test) ;
     double      getDistance(Data &x1,Data &x2);
-
+    virtual     void initModel();
+    virtual     void trainModel();
+    void        setOptimizer(Optimizer *p);
     void        loadTrainSet();
     void        loadTestSet();
     double      getTestError();
     double      getClassTestError();
     Dataset     *getTrainDataset();
+    virtual     void  testModel(double &trainError,double &testError,double &classError);
     virtual ~Model();
 };
 
