### 2024-10-21 52fafbf
improve fc model
