+++ rundata.sh
@@ -20,8 +20,8 @@ then
 elif [ $MODEL = "fc" ]
 then
 	CREATEPARAMS='\"--opt_method=Genetic\"\"--gen_maxiters=5\"\"--rbf_nodes=10\"\"--rbf_factor=2\"\"--opt_debug=no\"'
-	EVALPARAMS='\"--opt_method=Bfgs\"'
-	MODELPARAMS="--model_trainfile=$DATAPATH/$1.train --model_testfile=$DATAPATH/$1.test --fc_createmodel=rbf --fc_evaluatemodel=mlp --fc_popcount=500 --fc_popsize=200 --fc_popgens=200 --fc_features=2 --fc_createparams=$CREATEPARAMS --fc_evaluateparams=$EVALPARAMS"
+	EVALPARAMS='\"--opt_method=iPso\"\"--mlp_nodes=10\"\"--opt_debug=yes\"\"--mlp_initmethod=smallvalues\"\"--opt_termination=maxiters\"'
+	MODELPARAMS="--model_trainfile=$DATAPATH/$1.train --model_testfile=$DATAPATH/$1.test --fc_createmodel=rbf --fc_evaluatemodel=mlp --fc_popcount=200 --fc_popsize=200 --fc_popgens=10 --fc_features=2 --fc_createparams=$CREATEPARAMS --fc_evaluateparams=$EVALPARAMS"
 elif [ $MODEL = "rbf" ]
 then
 	MODELPARAMS="--rbf_nodes=10 --rbf_factor=8.0 --model_trainfile=$DATAPATH/$1.train --model_testfile=$DATAPATH/$1.test"
