### 2024-10-21 29e774e
add new woa method
