+++ moc_selectproblemdialog.cpp
@@ -1,0 +1,128 @@
+/****************************************************************************
+** Meta object code from reading C++ file 'selectproblemdialog.h'
+**
+** Created by: The Qt Meta Object Compiler version 67 (Qt 5.15.8)
+**
+** WARNING! All changes made in this file will be lost!
+*****************************************************************************/
+
+#include <memory>
+#include "XOPTIMUS/selectproblemdialog.h"
+#include <QtCore/qbytearray.h>
+#include <QtCore/qmetatype.h>
+#if !defined(Q_MOC_OUTPUT_REVISION)
+#error "The header file 'selectproblemdialog.h' doesn't include <QObject>."
+#elif Q_MOC_OUTPUT_REVISION != 67
+#error "This file was generated using the moc from 5.15.8. It"
+#error "cannot be used with the include files from this version of Qt."
+#error "(The moc has changed too much.)"
+#endif
+
+QT_BEGIN_MOC_NAMESPACE
+QT_WARNING_PUSH
+QT_WARNING_DISABLE_DEPRECATED
+struct qt_meta_stringdata_SelectProblemDialog_t {
+    QByteArrayData data[6];
+    char stringdata0[60];
+};
+#define QT_MOC_LITERAL(idx, ofs, len) \
+    Q_STATIC_BYTE_ARRAY_DATA_HEADER_INITIALIZER_WITH_OFFSET(len, \
+    qptrdiff(offsetof(qt_meta_stringdata_SelectProblemDialog_t, stringdata0) + ofs \
+        - idx * sizeof(QByteArrayData)) \
+    )
+static const qt_meta_stringdata_SelectProblemDialog_t qt_meta_stringdata_SelectProblemDialog = {
+    {
+QT_MOC_LITERAL(0, 0, 19), // "SelectProblemDialog"
+QT_MOC_LITERAL(1, 20, 6), // "okSlot"
+QT_MOC_LITERAL(2, 27, 0), // ""
+QT_MOC_LITERAL(3, 28, 10), // "cancelSlot"
+QT_MOC_LITERAL(4, 39, 15), // "comboSelectItem"
+QT_MOC_LITERAL(5, 55, 4) // "text"
+
+    },
+    "SelectProblemDialog\0okSlot\0\0cancelSlot\0"
+    "comboSelectItem\0text"
+};
+#undef QT_MOC_LITERAL
+
+static const uint qt_meta_data_SelectProblemDialog[] = {
+
+ // content:
+       8,       // revision
+       0,       // classname
+       0,    0, // classinfo
+       3,   14, // methods
+       0,    0, // properties
+       0,    0, // enums/sets
+       0,    0, // constructors
+       0,       // flags
+       0,       // signalCount
+
+ // slots: name, argc, parameters, tag, flags
+       1,    0,   29,    2, 0x0a /* Public */,
+       3,    0,   30,    2, 0x0a /* Public */,
+       4,    1,   31,    2, 0x0a /* Public */,
+
+ // slots: parameters
+    QMetaType::Void,
+    QMetaType::Void,
+    QMetaType::Void, QMetaType::QString,    5,
+
+       0        // eod
+};
+
+void SelectProblemDialog::qt_static_metacall(QObject *_o, QMetaObject::Call _c, int _id, void **_a)
+{
+    if (_c == QMetaObject::InvokeMetaMethod) {
+        auto *_t = static_cast<SelectProblemDialog *>(_o);
+        (void)_t;
+        switch (_id) {
+        case 0: _t->okSlot(); break;
+        case 1: _t->cancelSlot(); break;
[…]
