### 2024-10-22 09f8298
fixes in qvector
