### 2024-10-22 500a7da
fixes in qvector
