### 2024-10-22 3aa5f8e
data fitting fixes
