### 2024-10-23 c9753b6
fc fixes in rbf memory
