+++ moc_selectmethoddialog.cpp
@@ -1,128 +1,0 @@
-/****************************************************************************
-** Meta object code from reading C++ file 'selectmethoddialog.h'
-**
-** Created by: The Qt Meta Object Compiler version 67 (Qt 5.15.8)
-**
-** WARNING! All changes made in this file will be lost!
-*****************************************************************************/
-
-#include <memory>
-#include "XOPTIMUS/selectmethoddialog.h"
-#include <QtCore/qbytearray.h>
-#include <QtCore/qmetatype.h>
-#if !defined(Q_MOC_OUTPUT_REVISION)
-#error "The header file 'selectmethoddialog.h' doesn't include <QObject>."
-#elif Q_MOC_OUTPUT_REVISION != 67
-#error "This file was generated using the moc from 5.15.8. It"
-#error "cannot be used with the include files from this version of Qt."
-#error "(The moc has changed too much.)"
-#endif
-
-QT_BEGIN_MOC_NAMESPACE
-QT_WARNING_PUSH
-QT_WARNING_DISABLE_DEPRECATED
-struct qt_meta_stringdata_SelectMethodDialog_t {
-    QByteArrayData data[6];
-    char stringdata0[60];
-};
-#define QT_MOC_LITERAL(idx, ofs, len) \
-    Q_STATIC_BYTE_ARRAY_DATA_HEADER_INITIALIZER_WITH_OFFSET(len, \
-    qptrdiff(offsetof(qt_meta_stringdata_SelectMethodDialog_t, stringdata0) + ofs \
-        - idx * sizeof(QByteArrayData)) \
-    )
-static const qt_meta_stringdata_SelectMethodDialog_t qt_meta_stringdata_SelectMethodDialog = {
-    {
-QT_MOC_LITERAL(0, 0, 18), // "SelectMethodDialog"
-QT_MOC_LITERAL(1, 19, 16), // "onMethodSelected"
-QT_MOC_LITERAL(2, 36, 0), // ""
-QT_MOC_LITERAL(3, 37, 4), // "text"
-QT_MOC_LITERAL(4, 42, 6), // "okSlot"
-QT_MOC_LITERAL(5, 49, 10) // "cancelSlot"
-
-    },
-    "SelectMethodDialog\0onMethodSelected\0"
-    "\0text\0okSlot\0cancelSlot"
-};
-#undef QT_MOC_LITERAL
-
-static const uint qt_meta_data_SelectMethodDialog[] = {
-
- // content:
-       8,       // revision
-       0,       // classname
-       0,    0, // classinfo
-       3,   14, // methods
-       0,    0, // properties
-       0,    0, // enums/sets
-       0,    0, // constructors
-       0,       // flags
-       0,       // signalCount
-
- // slots: name, argc, parameters, tag, flags
-       1,    1,   29,    2, 0x0a /* Public */,
-       4,    0,   32,    2, 0x0a /* Public */,
-       5,    0,   33,    2, 0x0a /* Public */,
-
- // slots: parameters
-    QMetaType::Void, QMetaType::QString,    3,
-    QMetaType::Void,
-    QMetaType::Void,
-
-       0        // eod
-};
-
-void SelectMethodDialog::qt_static_metacall(QObject *_o, QMetaObject::Call _c, int _id, void **_a)
-{
-    if (_c == QMetaObject::InvokeMetaMethod) {
-        auto *_t = static_cast<SelectMethodDialog *>(_o);
-        (void)_t;
-        switch (_id) {
-        case 0: _t->onMethodSelected((*reinterpret_cast< const QString(*)>(_a[1]))); break;
-        case 1: _t->okSlot(); break;
[…]
