### 2024-11-07 d112253
hybrid doc
