+++ doc/hybridMethod/newHybrid/SUBMIT_HYBRID/hybridMethod.tex
@@ -603,8 +603,14 @@ results. The experiments were conducted on a system with an AMD Ryzen
 Additionally, the open-source optimizer \textquotedbl GlobalOptimus\textquotedbl{}
 was used, which is a fully developed optimizer and is available for
 distribution via the link: \url{http://www.github.com/itsoulos/GlobalOptimus}
-(accessed on 7 November 2024). In Table \ref{tab:mainTable}, the
-average function calls for each method and every objective function
+(accessed on 7 November 2024). This optimization software can be installed
+in any operating system (Linux, Windows, FreeBsd) when the Qt library
+is available. The software includes a variety of optimization methods
+as swll as some scripts to automate the optimization process under
+Windows or Linux. Also, the software provides a graphical user interface
+under the name Xoptimus and an installation file without dependencies
+for Windows system is also included. In Table \ref{tab:mainTable},
+the average function calls for each method and every objective function
 is presented. The following notation is used in the experimental tables:
 \begin{enumerate}
 \item The column FUNCTION represents the name of the used objective problem.
@@ -1378,8 +1384,8 @@ algorithms. Arabian Journal for Science and Engineering, 44, 2899-2923.
 multi-objective particle swarm optimization for large swarms and high
 dimensional problems. Parallel Computing, 92, 102589.
 
-\bibitem{holland}Holland, J. H. (1992). Genetic algorithms. Scientific
-american, 267(1), 66-73.
+\bibitem{holland}Eshelman, L. J. (2018). Genetic algorithms. In Evolutionary
+Computation 1 (pp. 102-118). CRC Press.
 
 \bibitem{ga_problem1}Hervis Santana, Y., Martinez Alonso, R., Guillen
 Nieto, G., Martens, L., Joseph, W., \& Plets, D. (2022). Indoor genetic
@@ -1396,10 +1402,9 @@ and Biotechnology, 9, 793782.
 potential of a high speed SI engine fueled with butanol--gasoline\LyXZeroWidthSpace{}
 blends. Energy Reports, 8, 69-80.
 
-\bibitem{key-29}Arifovic, J., \& Gencay, R. (2001). Using genetic
-algorithms to select architecture of a feedforward artificial neural
-network. Physica A: Statistical mechanics and its applications, 289(3-4),
-574-594.
+\bibitem{key-29}MirRokni, M. K. S. M. (2017). Applying genetic algorithm
+in architecture and neural network training. International Journal
+of Computer Science and Network Security IJCSN, 17(6), 118.
 
 \bibitem{de_symmetry1}Li, Y. H., Wang, J. Q., Wang, X. J., Zhao,
 Y. L., Lu, X. H., \& Liu, D. L. (2017). Community detection based
@@ -1435,9 +1440,9 @@ global optimization algorithm: particle swarm optimization in association
 with a genetic algorithm. Complex System Modelling and Control Through
 Intelligent Soft Computations, 45-86.
 
-\bibitem{armijo}Armijo, L. (1966). Minimization of functions having
-Lipschitz continuous first partial derivatives. Pacific Journal of
-mathematics, 16(1), 1-3.
+\bibitem{armijo}Huang, P., Huang, H. Z., \& Huang, T. (2019). A novel
+algorithm for structural reliability analysis based on finite step
+length and Armijo line search. Applied Sciences, 9(12), 2546.
 
 \bibitem[(2022)]{charilogis}Charilogis, V., \& Tsoulos, I. G. (2022).
 Toward an ideal particle swarm optimizer for multidimensional functions.
