+++ doc/hybridMethod/newHybrid/hybridMethod.tex
@@ -1,1496 +1,0 @@
-%% LyX 2.4.1 created this file.  For more info, see https://www.lyx.org/.
-%% Do not edit unless you really know what you are doing.
-\documentclass[journal,article,submit,pdftex,moreauthors]{mdpi}
-\usepackage[T1]{fontenc}
-\usepackage[utf8]{inputenc}
-\setcounter{secnumdepth}{1}
-\setcounter{tocdepth}{1}
-\usepackage{color}
-\usepackage{float}
-\usepackage{url}
-\usepackage{varwidth}
-\usepackage{graphicx}
-
-\makeatletter
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% LyX specific LaTeX commands.
-
-\Title{An innovative hybrid approach producing trial solutions for Global
-Optimization}
-
-\TitleCitation{An Innovative Hybrid Approach to Global Optimization}
-
-\Author{Vasileios Charilogis$^{1}$,Glykeria Kyrou$^{2}$, Ioannis G. Tsoulos$^{3,*}$,
-Anna Maria Gianni$^{4}$}
-
-\AuthorNames{V. Charilogis, Glykeria Kyrou, I.G. Tsoulos, A.M. Gianni}
-
-\AuthorCitation{Charilogis V.; Kyrou, G.;Tsoulos I.G.; Gianni A.M; }
-
-
-\address{$^{1}$\quad{}Department of Informatics and Telecommunications,
-University of Ioannina, Greece; v.charilog@uoi.gr\\
-$^{2}$\quad{}Department of Informatics and Telecommunications, University
-of Ioannina, Greece; g.kyrou@uoi.gr\\
-$^{3}$\quad{}Department of Informatics and Telecommunications, University
-of Ioannina, Greece; itsoulos@uoi.gr\\
-$^{4}$\quad{}Department of Informatics and Telecommunications, University
-of Ioannina, Greece; am.gianni@uoi.gr}
-
-
-\corres{Correspondence: itsoulos@uoi.gr;}
-
-
-\abstract{Global optimization is critical in engineering, computer science,
-and various industrial applications, as it aims to find optimal solutions
-for complex problems. The development of efficient algorithms has
-emerged from the need for optimization, with each algorithm offering
-specific advantages and disadvantages. An effective approach to solving
-complex problems is the hybrid method, which combines established
-global optimization algorithms. This paper presents a hybrid global
-optimization method, which produces trial solutions for an objective
-problem, utilizing Genetic Algorithm genetic operators, as well as
-solutions obtained through a linear search process. Then, the generated
-solutions are used to form new test solutions, by applying Differential
-Evolution techniques. These operations are based on samples derived
-either from internal line searches or genetically modified samples
-in specific subsets of Euclidean space. Additionally, other relevant
-approaches are explored to enhance the method's efficiency. The new
-method was applied on a wide series of benchmark problems from the
-recent literature and comparison was made against other established
-methods of Global Optimization. }
-
-
-\keyword{Optimization; Differential evolution; Genetic algorithm; Line search;
-Evolutionary techniques; Stochastic methods; Hybrid methods.}
-
-\newcommand*\LyXZeroWidthSpace{\hspace{0pt}}
-%% Because html converters don't know tabularnewline
-\providecommand{\tabularnewline}{\\}
-%% Variable width box for table cells
-\newenvironment{cellvarwidth}[1][t]
-    {\begin{varwidth}[#1]{\linewidth}}
-    {\@finalstrut\@arstrutbox\end{varwidth}}
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% User specified LaTeX commands.
-%  LaTeX support: latex@mdpi.com 
[…]
