### 2024-11-12 a6313b0
add JDE method
