### 2024-11-13 #7352 @PaulDebus
Hey,

I can give it a try. I am not an expert in C++ (though have some knowledge) and would need until the end of next week. If this is fine, I can take it on.
