### 2024-11-21 738059e
add sumfitness termination rule
