### 2024-11-26 370148c
add brenner potential
