+++ PROBLEMS/E_ev.cpp
@@ -1,0 +1,175 @@
+#include "E_ev.h"
+
+E_ev::E_ev()
+    : Problem(1) {}
+
+void E_ev::init(QJsonObject &params) {
+    int n = params["opt_dimension"].toString().toInt();
+    setDimension(n);
+    Data l, r;
+    l.resize(n);
+    r.resize(n);
+    for (int i = 0; i < n; i++) {
+        l[i] = -dimension;
+        r[i] = dimension;
+    }
+    setLeftMargin(l);
+    setRightMargin(r);
+}
+
+double E_ev::distance(const Data &x, size_t i, size_t j) {
+    double dx = x[i * 3] - x[j * 3];
+    double dy = x[i * 3 + 1] - x[j * 3 + 1];
+    double dz = x[i * 3 + 2] - x[j * 3 + 2];
+    return sqrt(dx * dx + dy * dy + dz * dz);
+}
+//Eq 4
+double E_ev::fij(double rij) {
+    const double R1 = 1.7;
+    const double R2 = 2.0;
+    if (rij < R1) {
+        return 1.0;
+    } else if (rij > R2) {
+        return 0.0;
+    }
+    double t = (rij - R1) * M_PI / (R2 - R1);
+    return 0.5 * (1 + cos(t));
+}
+//Eq 2
+double E_ev::VR(double rij) {
+    const double d = 6.325;
+    const double S = 1.29;
+    const double b = 1.5;
+    double e = exp(-sqrt(2 * S) * b * (rij - 1));
+    return fij(rij) * (d / (S - 1)) * e;
+}
+//Eq 3
+double E_ev::VA(double rij) {
+    const double d = 6.325;
+    const double S = 1.29;
+    const double b = 1.5;
+    double e = exp(-sqrt(2 / S) * b * (rij - 1));
+    return fij(rij) * (d * S / (S - 1)) * e;
+}
+
+double E_ev::calcNeighbor(const Data &x, size_t index) {
+    double sum = 0.0;
+    size_t n = dimension / 3;
+    for (size_t k = 0; k < n; ++k) {
+        double rik = distance(x, index, k);
+        sum += fij(rik);
+    }
+    return sum;
+}
+//Eq 8
+double E_ev::Ni(const Data &x, size_t i) {
+    return calcNeighbor(x, i);
+}
+//Eq 8
+double E_ev::Nj(const Data &x, size_t j) {
+    return calcNeighbor(x, j);
+}
+//Eq 9
+double E_ev::NijConj(const Data &x, size_t i, size_t j) {
+    double sum = 0.0;
+    size_t n = dimension / 3;
+
+    for (size_t k = 0; k < n; ++k) {
+        if (k != i && k != j) {
+            double rik = distance(x, i, k);
+            sum += fij(rik) * F(x, i, k);
+        }
+    }
+    for (size_t l = 0; l < n; ++l) {
+        if (l != j) {
+            double rjl = distance(x, j, l);
+            sum += fij(rjl) * F(x, j, l);
+        }
+    }
+    return sum;
+}
+//Eq 10
+double E_ev::F(const Data &x, size_t  i, size_t k) {
+    double xik = Ni(x, k);
+    if (xik > 2.0 && xik < 3.0) {
+        return 0.5 * (1 + cos(M_PI * (xik - 2)));
+    } else if (xik <= 2.0) {
+        return 1.0;
+    } else {
+        return 0.0;
+    }
+}
+//Eq 5,6
+double E_ev::Bij(const Data &x, size_t i, size_t j) {
+    double sum = 0.0;
+    size_t n = dimension / 3;
+
+    for (size_t k = 0; k < n; ++k) {
[…]
