+++ Makefile
@@ -103,6 +103,7 @@ SOURCES       = GE/cprogram.cc \
 		OPTIMUS/doublebox.cpp \
 		OPTIMUS/similarity.cpp \
 		PROBLEMS/Ackley.cpp \
+		PROBLEMS/E_ev.cpp \
 		PROBLEMS/Extended_f10.cpp \
 		PROBLEMS/Sphere.cpp \
 		PROBLEMS/bf1.cpp \
@@ -227,6 +228,7 @@ OBJECTS       = cprogram.o \
 		doublebox.o \
 		similarity.o \
 		Ackley.o \
+		E_ev.o \
 		Extended_f10.o \
 		Sphere.o \
 		bf1.o \
@@ -408,6 +410,7 @@ DIST          = /usr/lib/x86_64-linux-gnu/qt5/mkspecs/features/spec_pre.prf \
 		OPTIMUS/methodloader.h \
 		OPTIMUS/problemloader.h \
 		PROBLEMS/Ackley.h \
+		PROBLEMS/E_ev.h \
 		PROBLEMS/Extended_f10.h \
 		PROBLEMS/Sphere.h \
 		PROBLEMS/bf3.h \
@@ -560,6 +563,7 @@ DIST          = /usr/lib/x86_64-linux-gnu/qt5/mkspecs/features/spec_pre.prf \
 		OPTIMUS/doublebox.cpp \
 		OPTIMUS/similarity.cpp \
 		PROBLEMS/Ackley.cpp \
+		PROBLEMS/E_ev.cpp \
 		PROBLEMS/Extended_f10.cpp \
 		PROBLEMS/Sphere.cpp \
 		PROBLEMS/bf1.cpp \
@@ -834,8 +838,8 @@ distdir: FORCE
 	@test -d $(DISTDIR) || mkdir -p $(DISTDIR)
 	$(COPY_FILE) --parents $(DIST) $(DISTDIR)/
 	$(COPY_FILE) --parents /usr/lib/x86_64-linux-gnu/qt5/mkspecs/features/data/dummy.cpp $(DISTDIR)/
[…]
