### 2024-11-26 b796122
improve speed of DE
