+++ PROBLEMS/f13.h
@@ -8,6 +8,7 @@ public:
     f13 ();
     double funmin(Data &x);
     Data gradient(Data &x);
+    void init(QJsonObject &params);
 };
 
 #endif // F13_H
