+++ PROBLEMS/f16.h
@@ -8,6 +8,7 @@ public:
     f16 ();
     double funmin(Data &x);
     Data gradient(Data &x);
+    void init(QJsonObject &params);
 };
 
 #endif // F16_H
