+++ PROBLEMS/f9.h
@@ -8,6 +8,7 @@ public:
     f9 ();
     double funmin(Data &x);
     Data gradient(Data &x);
+    void init(QJsonObject &params);
 
 };
 
