+++ PROBLEMS/f1.h
@@ -8,6 +8,7 @@ public:
     F1 ();
     double funmin(Data &x);
     Data gradient(Data &x);
+    void init(QJsonObject &params);
 
 };
 
