+++ PROBLEMS/f18.h
@@ -8,6 +8,7 @@ public:
     f18 ();
     double funmin(Data &x);
     Data gradient(Data &x);
+    void init(QJsonObject &params);
 };
 
 #endif // F18_H

[…]
