### 2024-12-05 3fddace
de fixes
