+++ PROBLEMS/f3.h
@@ -2,10 +2,10 @@
 #define F3_H
 # include <OPTIMUS/problem.h>
 
-class F3 : public Problem
+class f3 : public Problem
 {
 public:
-    F3 ();
+    f3 ();
     double funmin(Data &x);
     Data gradient(Data &x);
     void init(QJsonObject &params);
