+++ PROBLEMS/f5.h
@@ -1,11 +1,11 @@
-#ifndef F5_H
-#define F5_H
+#ifndef f5_H
+#define f5_H
 # include <OPTIMUS/problem.h>
 
-class F5 : public Problem
+class f5 : public Problem
 {
 public:
-    F5 ();
+    f5 ();
     double funmin(Data &x);
     Data gradient(Data &x);
     void init(QJsonObject &params);

[…]
