+++ PROBLEMS/schwefel221.h
@@ -7,7 +7,7 @@ public:
     schwefel221 ();
     double funmin(Data &x);
     Data gradient(Data &x);
-
+void init(QJsonObject &params);
 };
 #endif// SCHWEFEL221_H
 
