+++ runfunmin.sh
@@ -24,7 +24,9 @@ then
 ##de_fselection:  The selection of differential weight. Values: number, ali, random, adaptive, migrant
 ##de_localsearch: Enable or disable the incorporation of local search at every iteration. Values: yes,no
 ##de_selection:   The selection method used in every iteration. Values: random, tournament
-	METHODPARAMS="--de_np=200 --de_maxiters=200 --de_selection=random --de_fselection=random --de_lrate=0.01 --opt_termination=sumfitness"
+	METHODPARAMS="--de_np=200 --de_maxiters=200 --de_selection=tournament --de_fselection=migrant --de_lrate=0.01 --opt_termination=sumfitness"
+
+
 elif [ $METHOD = "NelderMead" ]
 then
 #nm_population: number of items in simplex
