+++ runfunmin.sh
@@ -3,7 +3,7 @@ METHOD=DifferentialEvolution
 #Available local search methods: bfgs, lbfgs, gradient, nelderMead, adam
 LOCALSEARCH=bfgs
 #Available samplers: uniform, mlp, rbf, maxwell, triangular, kmeans
-SAMPLER=uniform
+SAMPLER=kmeans
 #Available stopping rules: maxiters, doublebox, similarity, mean, all, sumfitness (only for DE)
 TERMINATION=sumfitness
 ### GLOBAL PARAMS
@@ -24,7 +24,7 @@ then
 ##de_fselection:  The selection of differential weight. Values: number, ali, random, adaptive, migrant
 ##de_localsearch: Enable or disable the incorporation of local search at every iteration. Values: yes,no
 ##de_selection:   The selection method used in every iteration. Values: random, tournament
-	METHODPARAMS="--de_np=200 --de_maxiters=200 --de_selection=random --de_fselection=number --de_lrate=0.01 --opt_termination=sumfitness"
+	METHODPARAMS="--de_np=200 --de_maxiters=500 --de_selection=tournament --de_fselection=random --de_lrate=0.005 --opt_termination=sumfitness"
 
 
 elif [ $METHOD = "NelderMead" ]
