### 2024-12-11 #7352 @nicoguaro
@diehlpk, I will be on break next month. I could start a review in January. Also, it is been a while since I last used C++ so it might take me some time.
