+++ doc/LARGE_SCALE/SUBMIT_LARGE/large_scale.tex
@@ -1,0 +1,1821 @@
+%% LyX 2.3.7 created this file.  For more info, see http://www.lyx.org/.
+%% Do not edit unless you really know what you are doing.
+\documentclass[journal,article,submit,pdftex,moreauthors]{Definitions/mdpi}
+\usepackage[utf8]{inputenc}
+\usepackage{float}
+\usepackage{url}
+\usepackage{amsmath}
+\usepackage{graphicx}
+
+\makeatletter
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% LyX specific LaTeX commands.
+
+\Title{A novel method for large scale optimization problems, based on Differential
+Evolution}
+
+\TitleCitation{A novel method for large scale optimization problems, based on Differential
+Evolution}
+
+\Author{Glykeria Kyrou$^{1}$, Vasileios Charilogis$^{2}$ and Ioannis G.
+Tsoulos$^{3,*}$}
+
+\AuthorNames{Glykeria Kyrou, Vasileios Charilogis and Ioannis G. Tsoulos }
+
+\AuthorCitation{Kyrou, G.; Charilogis, V.; Tsoulos, I.G. }
+
+
+\address{$^{1}$\quad{}Department of Informatics and Telecommunications,
+University of Ioannina, 47150 Kostaki Artas, Greece; g.kyrou@uoi.gr\\
+$^{2}$\quad{}Department of Informatics and Telecommunications, University
+of Ioannina, Greece; v.charilog@uoi.gr\\
+$^{3}\quad$Department of Informatics and Telecommunications, University
+of Ioannina, 47150 Kostaki Artas, Greece;itsoulos@uoi.gr}
+
+
+\corres{Correspondence: itsoulos@uoi.gr}
+
+
+\abstract{Global optimization is fundamental to engineering and computer science
+as it seeks to find better solutions to both simple and complex problems.
+It aims to find the most effective and efficient solution to any problem.
+In this paper we present a variation of the differential evolution
+algorithm for large-scale Global Optimization problems. Differential
+Evolution (DE) is a universal optimization algorithm that is applied
+to many practical engineering topics. The DE algorithm is a population-based
+algorithm like genetic algorithms and uses similar operators such
+as: crossover, mutation and selection.In this work, a series of modifications
+are proposed that aim to improve the reliability and speed of the
+above technique. The new method was tested on a series of large-scale
+problems and compared with other global optimization techniques with
+promising results. More specifically, the proposed algorithm has been
+evaluated by typical high-dimensional numerical optimization problems.
+The functions used are from the CEC-2010 competition for Large-Scale
+Global Optimization problems.}
+
+
+\keyword{Optimization; Differential evolution; Evolutionary techniques; Stochastic
+methods; Large-Scale problems}
+
+\DeclareTextSymbolDefault{\textquotedbl}{T1}
+%% Because html converters don't know tabularnewline
+\providecommand{\tabularnewline}{\\}
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% User specified LaTeX commands.
+%  LaTeX support: latex@mdpi.com 
+%  For support, please attach all files needed for compiling as well as the log file, and specify your operating system, LaTeX version, and LaTeX editor.
+
+%=================================================================
+
+
+% For posting an early version of this manuscript as a preprint, you may use "preprints" as the journal and change "submit" to "accept". The document class line would be, e.g., \documentclass[preprints,article,accept,moreauthors,pdftex]{mdpi}. This is especially recommended for submission to arXiv, where line numbers should be removed before posting. For preprints.org, the editorial staff will make this change immediately prior to posting.
+
+%--------------------
[…]
