### 2024-12-15 d22eece
add de method
