### 2024-12-18 4fef02c
De add neural sampling
