### 2024-12-19 6f2e2fa
fix mlp init
