### 2025-01-02 476c398
rule model fixes
