+++ MLMODELS/rulemodel.cpp
@@ -42,7 +42,7 @@ void        RuleModel::trainModel()
 
  pop->setSelectionRate(getParam("rule_popsrate").getValue().toDouble());
  pop->setMutationRate(getParam("rule_popmrate").getValue().toDouble());
- pop->setLocalMethod(GELOCAL_SIMAN);
+ pop->setLocalMethod(GELOCAL_NONE);
 
 
  int gens = getParam("rule_popgens").getValue().toInt();
@@ -56,7 +56,13 @@ void        RuleModel::trainModel()
  for(int g=1;g<=gens;g++)
  {
      pop->nextGeneration();
-     printf(" generation = %d best value= %20.10lg\n",g,pop->getBestFitness());
+     int redo =0;
+     vector<int> genome = pop->getBestGenome();
+
+     if(g%50==0)
+     printf(" generation = %d best expression = %s best value= %20.10lg\n",g,
+            program->printRandomProgram(genome,redo).c_str(),
+            pop->getBestFitness());
      if(g%LI==0)
      {
          for(int i=1;i<=LC;i++)
