### 2025-01-03 ecc9f60
rule model improvements
