### 2025-01-03 0fe82e3
rule model fixes
