### 2025-01-06 itsoulos/GlobalOptimus#1 @PaulDebus
[JOSS] Review @PaulDebus
Hey,

I am one of the reviewers from the JOSS Journal https://github.com/openjournals/joss-reviews/issues/7584
and looking into your paper.

First of all: Thanks for submitting the paper, is looks interesting. I hope we can work together on making this a great contribution.

I will use this issue here for minor things and general discussions, and will open dedicated issues for bigger things (or those that do block the completion of the review). Those extra issues will also start with `[JOSS]` in the title.

Here is a list of those extra issues to keep track:

- https://github.com/itsoulos/GlobalOptimus/issues/2
- https://github.com/itsoulos/GlobalOptimus/issues/3
- https://github.com/itsoulos/GlobalOptimus/issues/5
- https://github.com/itsoulos/GlobalOptimus/issues/6
- https://github.com/itsoulos/GlobalOptimus/issues/8
- https://github.com/itsoulos/GlobalOptimus/issues/10
- https://github.com/itsoulos/GlobalOptimus/issues/11
- https://github.com/itsoulos/GlobalOptimus/issues/14
- https://github.com/itsoulos/GlobalOptimus/issues/15
- https://github.com/itsoulos/GlobalOptimus/issues/17

I would like to begin with a quick question: If I find minor issues, for example a bad formulation in the README file, would you like me to open a pull request or do you prefer fixing those yourself?

Best
Paul
