### 2025-01-06 #7584 @PaulDebus
I have started my review, but got stuck at the first hurdle: Making the software run.

The central issue for my review is this one: https://github.com/itsoulos/GlobalOptimus/issues/1
It should be considered blocking for the completion of the review. I open dedicated issues for larger topics in the software repository.
