### 2025-01-06 itsoulos/GlobalOptimus#1 @itsoulos
Yes of course
