### 2025-01-06 3e8a7d7
GE improvements
