+++ GE/population.cc
@@ -300,8 +300,9 @@ void	Population::nextGeneration()
 	calcFitnessArray();
 	
 	
-    const int mod=5;//localsearch_generations;
-    const int count=5;//localsearch_items;
+	
+    const int mod=localsearch_generations;
+    const int count=localsearch_items;
 	if((generation+1) % mod==0) 
 	{
 		for(int i=0;i<count;i++)
