### 2025-01-06 23a7e1c
Update README.md
Extend the installation docs for Linux and small formatting fixes
