### 2025-01-06 itsoulos/GlobalOptimus#6 @PaulDebus
[JOSS] Testing
The JOSS requires the software to have some kind of testing set up. I could not find it for this package.

I think an optimization package can be tested very elegantly by demonstrating that the optimal solutions can be found for well known and studied problems. The problems provided in this package maybe already provide this, as the output of the run contains the `Percent Minimum Found` line.

If this is the case, then a simple section in the README file would be enough to explain how to perform the testing.

It would be even better if there was an automated way to run the optimizations on all available problems with known optima and report if those can be found. This could also serve as a kind of unit test during development.
