### 2025-01-06 itsoulos/GlobalOptimus#6 @PaulDebus
This issue is blocking for the review: https://github.com/itsoulos/GlobalOptimus/issues/1
