### 2025-01-06 itsoulos/GlobalOptimus#6 @itsoulos
To test the software you should perform a test involving any test function and observe the statistics at the end of the output. These statistics include: average function calls, percentage of global minimum found and standard deviation of the calls.
