### 2025-01-06 itsoulos/GlobalOptimus#1 @PaulDebus
Could you please check the "Documentation" section of my review checklist in the main issue https://github.com/openjournals/joss-reviews/issues/7584

I think these points are not yet well met with the current documentation:
 - A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
- Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
- Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
- Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
- Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
- Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Especially the example usage and functionality documentation seem to be missing. I will add specific issues for this later on, as I think those points are crucial
