### 2025-01-06 itsoulos/GlobalOptimus#6 @PaulDebus
Is there a way to automate this? Especially for problems with known optima, this can be very elegant.

Also, please add a corresponding section to the README on how to test the quality of the found solutions. This would even be a good place to show some statistics of how close the found optima are to the known optima
