+++ paper/paper.bib
@@ -1,0 +1,243 @@
+@article{pant2020differential_2020,
+  title={Differential Evolution: A review of more than two decades of research},
+  author={Pant, Millie and Zaheer, Hira and Garcia-Hernandez, Laura and Abraham, Ajith and others},
+  journal={Engineering Applications of Artificial Intelligence},
+  volume={90},
+  pages={103479},
+  year={2020},
+  doi={10.1016/j.engappai.2020.103479},
+  publisher={Elsevier}
+}
+
+@article{charilogis2023parallel_2023,
+  title={A parallel implementation of the differential evolution method},
+  author={Charilogis, Vasileios and Tsoulos, Ioannis G},
+  journal={Analytics},
+  volume={2},
+  number={1},
+  pages={17--30},
+  year={2023},
+  doi={10.3390/analytics2010002},
+  publisher={MDPI}
+}
+
+@article{tsoulos2008modifications_2008,
+  title={Modifications of real code genetic algorithm for global optimization},
+  author={Tsoulos, Ioannis G},
+  journal={Applied Mathematics and Computation},
+  volume={203},
+  number={2},
+  pages={598--607},
+  year={2008},
+  doi={10.1016/j.amc.2008.05.005},
+  publisher={Elsevier}
+}
+
+@article{charilogis2022toward_2022,
+  title={Toward an ideal particle swarm optimizer for multidimensional functions},
+  author={Charilogis, Vasileios and Tsoulos, Ioannis G},
+  journal={Information},
+  volume={13},
+  number={5},
+  pages={217},
+  year={2022},
+  doi={10.3390/info13050217},
+  publisher={MDPI}
+}
+
+@article{tsoulos2023neuralminimizer_2023,
+  title={NeuralMinimizer: A Novel Method for Global Optimization},
+  author={Tsoulos, Ioannis G and Tzallas, Alexandros and Karvounis, Evangelos and Tsalikakis, Dimitrios},
+  journal={Information},
+  volume={14},
+  number={2},
+  pages={66},
+  year={2023},
+  doi={10.3390/info14020066 },
+  publisher={MDPI}
+}
+
+@article{charilogis2023improved_2023,
+  title={An improved parallel particle swarm optimization},
+  author={Charilogis, Vasileios and Tsoulos, Ioannis G and Tzallas, Alexandros},
+  journal={SN Computer Science},
+  volume={4},
+  number={6},
+  pages={766},
+  year={2023},
+  doi={10.1007/s42979-023-02227-9},
+  publisher={Springer}
+}
+
+@article{kirkpatrick1983optimization_1983,
+  title={Optimization by simulated annealing},
+  author={Kirkpatrick, Scott and Gelatt Jr, C Daniel and Vecchi, Mario P},
+  journal={science},
+  volume={220},
+  number={4598},
+  pages={671--680},
+  year={1983},
+  doi={10.1126/science.220.4598.671},
+  publisher={American association for the advancement of science}
+}
+
+@article{kyrou2024eofa_2024,
+  title={EOFA: An Extended Version of the Optimal Foraging Algorithm for Global Optimization Problems},
+  author={Kyrou, Glykeria and Charilogis, Vasileios and Tsoulos, Ioannis G},
+  journal={Computation},
+  volume={12},
+  number={8},
+  pages={158},
+  year={2024},
+  doi={10.3390/computation12080158 },
+  publisher={MDPI}
+}
+
+@article{powell1989tolerant_1989,
+  title={A tolerant algorithm for linearly constrained optimization calculations},
+  author={Powell, MJD},
+  journal={Mathematical Programming},
+  volume={45},
[…]
