### 2025-01-09 itsoulos/GlobalOptimus#6 @PaulDebus
I think such a test should also show the quality of the results and demonstrate that they are correct. This is also a requirement of the JOSS reviewers checklist.  It specifically asks for automated tests.

Could you please look into that?
