### 2025-01-10 aebff44
add rastrigin description
