+++ OPTIMUS/statistics.cpp
@@ -59,7 +59,7 @@ void Statistics::printStatistics()
     printf("Average Function calls: %20.2lf\n",totalFcalls*1.0/n);
     printf("Minimum Function Value: %20.10lg\n",miny);
     printf("Percent Minimum  Found: %20.2lf%%\n",timesMinimumFound*100.0/n);
-    printf("Standard Deviation of Function Calls: %20.10lg\n", std_deviation_fcalls);
+    //printf("Standard Deviation of Function Calls: %20.10lg\n", std_deviation_fcalls);
 
     QStringList uniqueNames;
     for(int i=0;i<functionMetric.size();i++)
