### 2025-01-10 itsoulos/GlobalOptimus#6 @itsoulos
We have modified the Testing section in the README.md and now it contains the description of the parameters passed to GlobalOptimus executable as well as some example runs used to test the effectiveness of the Genetic Algorithm optimizer when it is applied a test function.
