### 2025-01-12 d0fa143
add mutate method
