### 2025-01-13 8017a61
nnc improvements
