### 2025-01-15 itsoulos/GlobalOptimus#6 @PaulDebus
I think this description helps to use the software and makes the importance of the parameters clear.

For automated testing, I still see some need for improvement. As I understand it, automated testing could be simplified to:
1. We know what our software is supposed to do
2. We run the software and record what it does
3. We compare the two and look if it works as intended

In the context of an optimizer, my expectation would be:
1. We have one (or multiple) established optimization problems with known optima
2. We run our optimizer on these problems and record the reported optima
3. We compare the target optima to the found optima and can analyze the performance

I think this could be rather easy to implement and would provide immense benefit, as we then would have an estimation of the performance of the optimizer.

@nicoguaro can we get your perspective here? I do not want to be unreasonable, but I believe some form of testing is required by the JOSS
