### 2025-01-15 itsoulos/GlobalOptimus#6 @itsoulos
@PaulDebus  the software prints at the end of the optimization the located global optimum for every run (and you can compare it with other software) as well as some needed information for optimization, such as:
1) Average function calls
2) Success rate (percentage rate where the global minimum was discovered)

Check the README.md for details about it
