### 2025-01-15 itsoulos/GlobalOptimus#1 @PaulDebus
I made a pull request (https://github.com/itsoulos/GlobalOptimus/pull/9) to change some wording and maybe make a bit of space for more words.

The JOSS requirements need a quick overview over the state of the field. So maybe you can look to include a quick mention of similar solutions and differences to Optimus?
