### 2025-01-15 bcaf99d
add more details for user defined problems
