### 2025-01-15 d06e991
add more details for user defined problems
