### 2025-01-15 06c81d5
add more details for user defined problems
