### 2025-01-15 8317a8b
Update README.md
add more clarification about recompiling the software after adding a custom problem instance
