### 2025-01-15 e305daa
add more details for user defined problems
