### 2025-01-16 bba5c52
nnc improvements
