+++ MLMODELS/nncmodel.cpp
@@ -1,6 +1,5 @@
 #include "nncmodel.h"
 
-
 NNCModel::NNCModel()
 {
     program = NULL;
@@ -17,6 +16,10 @@ NNCModel::NNCModel()
     addParam(Parameter("nnc_lsearchrate",0.0,0.0,1.0,"Local search rate for nnc"));
     addParam(Parameter("nnc_crossitems",100,10,10000,"Crossover items"));
     addParam(Parameter("nnc_steadystate","no","Enable or disable steady state"));
+    addParam(Parameter("nnc_weightfactor",2.0,1.0,10.0,"The weight factor for local optimization"));
+    QStringList yesno;
+    yesno<<"no"<<"yes";
+    addParam(Parameter("nnc_enablebound",yesno[0],yesno,"Enable (yes) or disable (no) the bounding solution for MLP"));
     QStringList methods;
     methods<<"crossover"<<"mutate"<<"bfgs"<<"none";
     addParam(Parameter("nnc_lsearchmethod",methods[0],methods,"Available methods: crossover,mutate,siman,bfgs,none"));
@@ -76,14 +79,18 @@ void    NNCModel::trainModel()
         pop->setLocalMethod(GELOCAL_BFGS);
     else
         pop->setLocalMethod(GELOCAL_SIMAN);
+
+    /**
+     */
+    pop->setLocalMethod(GELOCAL_NONE);
     pop->setLocalSearchGenerations(LI);
     pop->setLocalSearchItems(LC);
     for(int g=1;g<=gens;g++)
     {
         pop->nextGeneration();
-        if(g%50==0)
+        if(g%10==0)
         printf(" generation = %d best value= %20.10lg\n",g,pop->getBestFitness());
-        /*if(g%LI==0)
+        if(g%LI==0)
             {
                 for(int i=1;i<=LC;i++)
                 {
@@ -91,50 +98,35 @@ void    NNCModel::trainModel()
 
                     localSearchItem(pos);
                 }
-        }*/
-	//if(g%50==0) localSearchItem(0);
-       }
+        }
 
+       }
     pop->evaluateBestFitness();
-   setParam("nnc_lsearchmethod","bfgs");
-   localSearchItem(0);
-   setParam("nnc_lsearchmethod",Lmethod);
-    if(Lmethod == "none")
-        pop->setLocalMethod(GELOCAL_NONE);
-    else
-        if(Lmethod == "crossover")
-        pop->setLocalMethod(GELOCAL_CROSSOVER);
-    else
-        if(Lmethod == "mutate")
-        pop->setLocalMethod(GELOCAL_MUTATE);
-    else
-        if(Lmethod == "bfgs")
-        pop->setLocalMethod(GELOCAL_BFGS);
-    else
-        pop->setLocalMethod(GELOCAL_SIMAN);
 }
 
 void        NNCModel::localSearchItem(int pos)
 {
-       QString Lmethod = getParam("nnc_lsearchmethod").getValue();
-       Lmethod = "bfgs";
+    Data w;
+    vector<int> xx;
+    double yy;
+    int redo = 0;
+    double wf = getParam("nnc_weightfactor").getValue().toDouble();
+ QString Lmethod = getParam("nnc_lsearchmethod").getValue();
+
        if(Lmethod == "random")
        {
         pop->localSearch(pos);
         return;
        }
-        vector<int> xx;
-        double yy;
+
         pop->getChromosome(pos,xx,yy);
-	vector<int> oldx = xx;
+        vector<int> oldx = xx;
         if(fabs(yy)>1e+10) return;
-        int redo = 0;
         string st = program->printRandomProgram(xx,redo);
         if(!program->Parse(st)) return;
         parser->makeVector(st);
         string pt = parser->print();
 
-       Data w;
         parser->getWeights(w);
[…]
