### 2025-01-16 2926851
nnc improvements
