### 2025-01-17 6c33e9c
nnc improvements
