### 2025-01-18 f93133f
remove files not needed from DataFitting.pro
