+++ Makefile
@@ -108,42 +108,7 @@ SOURCES       = GE/cprogram.cc \
 		OPTIMUS/collection.cpp \
 		OPTIMUS/doublebox.cpp \
 		OPTIMUS/similarity.cpp \
-		PROBLEMS/bf1.cpp \
-		PROBLEMS/bf2.cpp \
-		PROBLEMS/branin.cpp \
-		PROBLEMS/camel.cpp \
-		PROBLEMS/cm.cpp \
-		PROBLEMS/diffpower.cpp \
-		PROBLEMS/diracproblem.cpp \
-		PROBLEMS/easom.cpp \
-		PROBLEMS/elp.cpp \
-		PROBLEMS/exp.cpp \
-		PROBLEMS/fuch1.cpp \
-		PROBLEMS/fuchss.cpp \
-		PROBLEMS/gkls.cpp \
-		PROBLEMS/gkls2100.cpp \
-		PROBLEMS/gkls250.cpp \
-		PROBLEMS/E_ev.cpp \
-		PROBLEMS/gkls350.cpp \
-		PROBLEMS/goldstein.cpp \
-		PROBLEMS/griewank10.cpp \
-		PROBLEMS/griewank2.cpp \
-		PROBLEMS/hansen.cpp \
-		PROBLEMS/hartman3.cpp \
-		PROBLEMS/hartman6.cpp \
-		PROBLEMS/hess.cpp \
-		PROBLEMS/levy.cpp \
-		PROBLEMS/potential.cpp \
-		PROBLEMS/rosenbrock.cpp \
-		PROBLEMS/salkin.cpp \
-		PROBLEMS/shekel10.cpp \
-		PROBLEMS/shekel5.cpp \
-		PROBLEMS/gross.cpp \
-		PROBLEMS/shekel7.cpp \
-		PROBLEMS/sinu.cpp \
-		PROBLEMS/test30n.cpp \
 		MLMODELS/dataset.cpp \
-		PROBLEMS/userproblem.cpp \
 		SAMPLER/kmeanssampler.cpp \
 		SAMPLER/maxwellsampler.cpp \
 		SAMPLER/neuralsampler.cpp \
@@ -152,14 +117,10 @@ SOURCES       = GE/cprogram.cc \
 		SAMPLER/uniformsampler.cpp \
 		datafitting.cpp \
 		OPTIMUS/optimizer.cpp \
-		PROBLEMS/discus.cpp \
-		PROBLEMS/cigar.cpp \
 		OPTIMUS/parameter.cpp \
 		OPTIMUS/problem.cpp \
 		MLMODELS/converter.cc \
 		MLMODELS/neuralparser.cc \
-		PROBLEMS/rastriginproblem.cpp \
-		PROBLEMS/test2nproblem.cpp \
 		SAMPLER/problemsampler.cpp 
 OBJECTS       = cprogram.o \
 		doublestack.o \
@@ -217,42 +178,7 @@ OBJECTS       = cprogram.o \
 		collection.o \
 		doublebox.o \
 		similarity.o \
-		bf1.o \
-		bf2.o \
-		branin.o \
-		camel.o \
-		cm.o \
-		diffpower.o \
-		diracproblem.o \
-		easom.o \
-		elp.o \
-		exp.o \
-		fuch1.o \
-		fuchss.o \
-		gkls.o \
-		gkls2100.o \
-		gkls250.o \
-		E_ev.o \
-		gkls350.o \
-		goldstein.o \
-		griewank10.o \
-		griewank2.o \
-		hansen.o \
-		hartman3.o \
-		hartman6.o \
-		hess.o \
-		levy.o \
-		potential.o \
-		rosenbrock.o \
-		salkin.o \
-		shekel10.o \
[…]
