### 2025-01-20 itsoulos/GlobalOptimus#6 @itsoulos
@PaulDebus Is the above information sufficient to close this topic?
