### 2025-01-21 itsoulos/GlobalOptimus#6 @nicoguaro
@PaulDebus, I am just resuming my review duties. But after checking the repo and your comments, I agree with you. That is, it appears that the software lacks some form of automated testing.
