### 2025-01-21 itsoulos/GlobalOptimus#6 @itsoulos
@nicoguaro  the software displays at the end the percentage of global minima discovered in the independent runs, the average number of function calls as well as the report for every run. What else testing should we perform?
