### 2025-01-21 itsoulos/GlobalOptimus#6 @nicoguaro
You are referring to tests of the software execution. On the other hand, we are referring to software testing. A tool of test that checks different functions/classes/methods and aids in the development of the software.
