### 2025-01-21 itsoulos/GlobalOptimus#6 @itsoulos
The software consists of a series of optimization methods (global and local) and many problems from the relevant literature. I do not think that could be usable to add such a functionality in this software.
