### 2025-01-22 itsoulos/GlobalOptimus#6 @PaulDebus
I actually think it would be especially useful to have such information about the software. These software tests are used to assure the correct functioning of the software by analyzing the behavior.

If you take an optimization method from the literature, they probably report results on some problems they evaluated. To make sure that your implementation is correct, you can optimize the same problems with your implementation and test that you reproduce the original results.

From a users' perspective, this is really important: If I use this software to optimize some problem from my domain, I want to be able to trust those results. And this trust can be established by demonstrating that the optimizers work as described on the provided example problems.

As an example, lets take the Griewank equation. There is a known single global minimum at $x^*$, where the function has minimal value $f ( x^∗ ) = 0$. The test setup would be to run an optimization method on this problem and check, whether this known optimum is reached.
This can be automated, so for example, you could test every solver on every problem with a known optimum. Make this functionality available to the users, so they have a good understanding of the performance they can expect. This helps spot errors in the implementation and can support in selecting a suitable optimization method if you know the characteristics of your problem.

In mission-critical software, this testing can be further broken down to test every single module or function in the program by testing if the expected behavior is observed for known input. This could for example also be implemented for the optimization problems to make sure there are no errors. I think such detailed testing is nice to have, but not a necessity for this particular project.

But in my perspective, having these fundamental tests that the implemented optimization methods achieve the results from their original reports is important.
