### 2025-01-22 itsoulos/GlobalOptimus#6 @PaulDebus
Here is a Wikipedia article that might be helpful to find test functions with known optima: https://en.wikipedia.org/wiki/Test_functions_for_optimization
