### 2025-01-22 itsoulos/GlobalOptimus#1 @PaulDebus
Any updates to the above comments?
