### 2025-01-22 #7584 @PaulDebus
I answered @itsoulos in the issue I had created in the projects repo.

Our review is active and ongoing. I made a dedicated issue here: https://github.com/itsoulos/GlobalOptimus/issues/1
