### 2025-01-22 itsoulos/GlobalOptimus#1 @itsoulos
I have already answered to the most of them. For example, the installation procedure, the description of the test problems, the user defined problems etc. Is there any list of point to point questions to work with it ?
