### 2025-01-22 itsoulos/GlobalOptimus#6 @itsoulos
The distribution contains also the runfunmin.sh (for Linux environments) and the runfunmin.bat for windows environments to automate the testing procedure.
