### 2025-01-22 itsoulos/GlobalOptimus#1 @PaulDebus
I made two more issues and put them into the list at the top of the issue
