+++ Contributing.md
@@ -1,0 +1,49 @@
+Contributing
+
+Admin: Ioannis G. Tsoulos
+
+Thank you for your interest in contributing to our project! Contributions from the community are of great importance for its continuous improvement and development. Below you will find instructions on how you can help.
+
+
+Getting started  
+
+1. Read README.md: Learn about the project, its structure, and its purpose.  
+2. Set up your environment. 
+3. Look at the existing code: Familiarize yourself with the structure of the project and its logic.
+
+Submit Questions  
+
+- For general questions or clarifications, please open a new issue (Issue) on GitHub.  
+- Bug Questions:
+  - Describe the problem in detail.  
+  - Include information such as your operating system, compiler version, and problem steps.
+
+Error Reporting  
+
+If you find an error:  
+1. Open a new issue (Issue) on GitHub.  
+2. Include:  
+   - A brief description of the problem.  
+   - How to reproduce it (steps, sample code).  
+   - Any error messages.  
+3. If possible, suggest a possible solution.
+
+
+Contribution to the Code  
+
+1. Fork the repository: Create a clone to work on the code.  
+2. Make your changes:  
+   - Make sure your changes work correctly without causing new problems.  
+4. Submit a Pull Request (PR):  
+   - Clearly describe your changes.  
+   - Report related Issues or document why your change is needed.
+
+
+ Documentation  
+
+- Update the documentation for any new functionality or code change.  
+- Check that README.md and other documentation files are up to date.  
+
+
+
+Thank you for your contribution! With your help, we can make our project better!
