### 2025-02-06 itsoulos/GlobalOptimus#6 @itsoulos
The README.md has been updated and now it includes a short description of the command line  utility runfunmin.sh, that can be used to test every optimization method included in the software
