### 2025-02-06 itsoulos/GlobalOptimus#6 @itsoulos
I do not think that the software should focus on automatic tests, but if you have any suggestion please send it here or through email to me
