### 2025-02-06 itsoulos/GlobalOptimus#6 @PaulDebus
I can put them here, so we can easily discuss the idea.

For the preparation, two steps are required:
1. Extend the `Problem` class with two properties, the known optimal value and the location of that known optimal value. Those can be optional for problems where no known optimum exists.
2. Add those to all problem implementations that have known optima (see the wikipedia link above)

For the use of this, I see different ways:

The known optimum can be integrated into the existing `GlobalOptimus` executable, such that after running, it checks if the problem class has a known optimum, reports if that optimum was found and changes the exit code of the program to `0` if the optimum was found and `1` if it was not found. This is similar to what it already does and requires only small changes.

You could also create a new executable especially for testing, where you can specify problem(s) and method(s) to test. So for example, you can set it up to run all methods on a specific problem, run a method on all problems, run all methods on all problems, just one method and one problem.... and always check if the known optimum was found. This is a bit of programming effort and a lot of computation, but this would allow you to have automated testing of the functionality, to make sure you do not introduce bugs while working on the program
