### 2025-02-06 itsoulos/GlobalOptimus#6 @nicoguaro
> Looks good. But it still does not address my major point in this issue: Automated testing of the software with regard to performance and correctness.
>
> I think this is really important and the way I understand the JOSS review guidelines also required. Please consider adding some automated tests. We can also discuss how this could be done, I have some ideas on how to integrate automated testing.
>
> If you absolutely do not want to add automated testing, I would refer the decision about this aspect to the JOSS editors and ask them to explain the requirement for automated tests.

I also consider that this is one relevant point regarding JOSS submissions.
