### 2025-02-06 itsoulos/GlobalOptimus#6 @nicoguaro
> I do not think that the software should focus on automatic tests, but if you have any suggestion please send it here or through email to me

We are not suggesting that automatic testing should be the focus of the software. It is a good practice from a software engineering perspective. You might want to check the following papers.

- Wilson, G., Aruliah, D. A., Brown, C. T., Hong, N. P. C., Davis, M., Guy, R. T., Haddock, S. H. D., Huff, K. D., Mitchell, I. M., Plumbley, M. D., Waugh, B., White, E. P., & Wilson, P. (2014). Best Practices for Scientific Computing. PLOS Biology, 12(1), e1001745. https://doi.org/10.1371/journal.pbio.1001745

- Wilson, G., Bryan, J., Cranston, K., Kitzes, J., Nederbragt, L., & Teal, T. K. (2017). Good enough practices in scientific computing. PLOS Computational Biology, 13(6), e1005510. https://doi.org/10.1371/journal.pcbi.1005510
