### 2025-02-08 959f05c
add de local method in Genetic
