+++ doc/NNC_GA/SUBMIT_NNC_GA/nnc_ga.tex
@@ -1,0 +1,1694 @@
+%% LyX 2.3.7 created this file.  For more info, see http://www.lyx.org/.
+%% Do not edit unless you really know what you are doing.
+\documentclass[journal,article,submit,pdftex,moreauthors]{Definitions/mdpi}
+\usepackage[utf8]{inputenc}
+\usepackage{float}
+\usepackage{url}
+\usepackage{graphicx}
+
+\makeatletter
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% LyX specific LaTeX commands.
+
+\Title{Combining constructed artificial neural networks with parameter constraint
+techniques to achieve better generalization properties}
+
+\TitleCitation{Combining constructed artificial neural networks with parameter constraint
+techniques to achieve better generalization properties}
+
+\Author{Ioannis G. Tsoulos$^{1,*}$, Vasileios Charilogis$^{2}$, Dimitrios
+Tsalikakis$^{3}$}
+
+\AuthorNames{Ioannis G. Tsoulos, Vasileios Charilogis, Dimitrios Tsalikakis}
+
+\AuthorCitation{Tsoulos, I.G.; Charilogis, V.; Tsalikakis D.}
+
+
+\address{$^{1}$\quad{}Department of Informatics and Telecommunications,
+University of Ioannina, Greece;itsoulos@uoi.gr\\
+$^{2}$\quad{}Department of Informatics and Telecommunications, University
+of Ioannina, Greece; v.charilog@uoi.gr\\
+$^{3}\quad$Department of Engineering Informatics and Telecommunications,
+University of Western Macedonia, 50100 Kozani, Greece; tsalikakis@gmail.com}
+
+
+\corres{Correspondence: itsoulos@uoi.gr}
+
+
+\abstract{A machine learning technology that has been used in a wide range
+of everyday problems is the recently proposed technique of constructing
+artificial neural networks with the assistance of Grammatical Evolution.
+This technique was utilized in problems from the fields of physics,
+chemistry, medicine, etc., in which it had excellent performance compared
+to other machine learning models. A key feature of these machine learning
+models is that they can both construct the structure of an artificial
+neural network efficiently and identify optimal values of the parameters
+of this network at the same time. However, in many cases the optimization
+process can become trapped in local minima of the error function,
+resulting in the construction of the artificial neural network not
+being able to be completed satisfactorily, which will lead to poor
+results in terms of test error. In this work, the combination of the
+artificial neural network construction method with an adapted genetic
+algorithm is proposed, which preserves the structure of the neural
+network that has been constructed and, in addition, appropriately
+adapts the parameters of the artificial neural network in order to
+avoid overfitting problems. The proposed method was applied on a wide
+series of datasets from real world problems that include classification
+and regression problems and it was compared against traditional machine
+learning models and the results are reported.}
+
+
+\keyword{Grammatical Evolution; Genetic Programming; Neural networks; Local
+Optimization}
+
+\DeclareTextSymbolDefault{\textquotedbl}{T1}
+%% Because html converters don't know tabularnewline
+\providecommand{\tabularnewline}{\\}
+\floatstyle{ruled}
+\newfloat{algorithm}{tbp}{loa}
+\providecommand{\algorithmname}{Algorithm}
+\floatname{algorithm}{\protect\algorithmname}
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% Textclass specific LaTeX commands.
+\newenvironment{lyxcode}
+	{\par\begin{list}{}{
+		\setlength{\rightmargin}{\leftmargin}
+		\setlength{\listparindent}{0pt}% needed for AMS classes
+		\raggedright
[…]
