### 2025-02-11 itsoulos/GlobalOptimus#1 @diehlpk
@itsoulos how are things going?
