### 2025-02-12 #7584 @nicoguaro
@diehlpk, we are having a discussion regarding [testing](https://github.com/itsoulos/GlobalOptimus/issues/6)
