### 2025-02-12 itsoulos/GlobalOptimus#1 @itsoulos
@diehlpk  hello. We are finalizing now some tests on GlobalOptimus. Sorry about the delay, but we had to submit some articles these days with my team
