### 2025-02-12 14cac5e
add has known global minimum functions
