+++ OPTIMUS/problemloader.cpp
@@ -73,29 +73,69 @@ QString ProblemLoader::getProblemReport(QString selectedproblem,
 Problem *ProblemLoader::getProblemFromName(QString name)
 {
     Problem *myProblem=NULL;
+    Data xx;
+    double yy;
     if(name == "rastrigin")
+    {
         myProblem = new RastriginProblem();
+        xx.resize(2);
+        xx[0]=0;
+        xx[1]=1;
+        yy= -2.0;
+        myProblem->setKnownOptimum(yy,xx);
+    }
     else
     if(name == "ackley")
+    {
         myProblem = new Ackley();
+        xx.resize(2);
+        xx[0]=0;
+        xx[1]=0;
+        yy= 0.0;
+        myProblem->setKnownOptimum(yy,xx);
+    }
     else
     if(name == "bf3")
+    {
+
         myProblem = new bf3();
+        xx.resize(2);
+        xx[0]=0;
+        xx[1]=0;
+        yy= 0.0;
+        myProblem->setKnownOptimum(yy,xx);
+    }
     else
     if(name == "extendedf10")
+    {
+
         myProblem = new Extended_f10();
+        xx.resize(2);
+        xx[0]=0;
+        xx[1]=0;
+        yy= 0.0;
+        myProblem->setKnownOptimum(yy,xx);
+    }
     else
     if(name == "schwefel")
+    {
+
         myProblem = new schwefel();
+    }
     else
     if(name == "schwefel221")
-        myProblem = new schwefel221();
+    {
+       myProblem = new schwefel221();
+    }
     else
     if(name == "schwefel222")
         myProblem = new schwefel222();
     else
     if(name  == "schaffer")
+    {
+
         myProblem = new schaffer();
+    }
     else
     if(name == "f16")
         myProblem = new f16();
@@ -128,16 +168,45 @@ Problem *ProblemLoader::getProblemFromName(QString name)
         myProblem = new Test2nProblem();
     else
     if(name == "bf1")
+    {
+
         myProblem = new Bf1();
+        xx.resize(2);
+        xx[0]=0;
+        xx[1]=0;
+        yy= 0.0;
+        myProblem->setKnownOptimum(yy,xx);
+    }
     else
     if(name == "bf2")
+    {
         myProblem = new Bf2();
+        xx.resize(2);
+        xx[0]=0;
+        xx[1]=0;
+        yy= 0.0;
+        myProblem->setKnownOptimum(yy,xx);
+    }
     else
     if(name == "branin")
+    {
         myProblem = new Branin();
+        xx.resize(2);
+        xx[0]=-3.141593;
+        xx[1]=5.683451;
+        yy= 0.397887;
+        myProblem->setKnownOptimum(yy,xx);
+    }
     else
     if(name == "camel")
+    {
         myProblem = new Camel();
+        xx.resize(2);
+        xx[0]=-0.089842;
+        xx[1]=0.712656;
+        yy= -1.031628 ;
+        myProblem->setKnownOptimum(yy,xx);
+    }
     else
     if(name == "cm")
         myProblem = new Cm();
@@ -146,7 +215,14 @@ Problem *ProblemLoader::getProblemFromName(QString name)
         myProblem = new DiffPower();
     else
     if(name == "easom")
+    {
         myProblem = new Easom();
+        xx.resize(2);
+        xx[0]=-87.964594;
+        xx[1]= 85.636858;
+        yy= -1.0;
+        myProblem->setKnownOptimum(yy,xx);
+    }
     else
     if(name == "exp")
         myProblem = new Exp();
[…]
