### 2025-02-12 itsoulos/GlobalOptimus#6 @itsoulos
Following your suggestions we have added the following methods in Problem class:
 void setKnownOptimum(double value, const Data& location);
 bool hasOptimum() const;
 double getKnownOptimumValue() const;
 Data getKnownOptimumLocation() const;
These methods can be used when the user knows the expected global minimum and they can be used to test any optimization method for its efficiency. Also, we have added a subsection in README.md under the name "Comparing with known global optima" to describe the previous process.
