### 2025-02-14 itsoulos/GlobalOptimus#6 @PaulDebus
that sounds promising. i am really busy with my own paper deadline right now, but i will take a look next week ;)
