### 2025-02-17 itsoulos/GlobalOptimus#6 @PaulDebus
Ok, this looks really good. This brings an important component to validate the performance of the algorithms.

I have two suggestions to further improve the code and the paper:
1. You could build on these latest changes to include a testing procedure (either a new program or an option for the existing one) that evaluates all the algorithms on these problems and reports the results. This way, the performance over all problems can be tested and if any changes to the code change these results, a regression can easily be spotted.
2. Having established the performance of the algorithms on the different problems, you could report them in the README and the paper. This would give important information to the users about the reliability of the implementation. You could even compare your performance against the original implementations of the algorithms
