+++ METHODS/hybridmethod.cpp
@@ -302,7 +302,6 @@ void HybridMethod::done()
     after = std::chrono::system_clock::now();
     auto milliseconds = std::chrono::duration_cast<std::chrono::milliseconds>(after - before);
     auto ms = milliseconds.count();
-    std::cout << "Douration: " << (double)ms / 1000.0 << " sec" << std::endl;
 }
 
 HybridMethod::~HybridMethod()
