+++ METHODS/parallelpso.cpp
@@ -469,7 +469,6 @@ void ParallelPso::done()
     after = std::chrono::system_clock::now();
     auto milliseconds = std::chrono::duration_cast<std::chrono::milliseconds>(after - before);
     auto ms = milliseconds.count();
-//    std::cout << "Douration: " << (double)ms / 1000.0 << " sec" << std::endl;
 }
 
 double ParallelPso::fitness(Data &x)
