### 2025-02-19 itsoulos/GlobalOptimus#6 @itsoulos
Following your suggestions we have included the method  void runAllTests(int iterations)  in the main.cpp and we have also added the option --opt_enabletesting in the GlobalOptimus in order to apply a series of global optimization methods on a selected problem. Also, we have extended the README.md to describe this procedure and we have added a table with some results using this procedure in the section "Comparing with known global optima" of the README.md file.
