+++ paper/paper.md
@@ -40,6 +40,8 @@ The location of global minima is applied in a variety of problems derived from p
 # Statement of need
 <p style="text-align: justify;">
 This paper presents the `GlobalOptimus` software,  written in ANSI C++, which allows researchers wanting to solve optimization problems in their domain to formulate their optimization problems and apply different global optimization methods to identify the global minimum of the problem. The main features of the proposed software are: a) Coding of the objective problem in a high level language such as ANSI C++ b) Incorporation of many global optimization techniques to solve the objective problem c) Parameterization of global optimization methods using user-defined parameters d) Usage of a GUI application to control the optimization strategy.
+<br>
+Some of the frequently used optimization tools are MATLAB Optimization Toolbox, SciPy optimization module and Gurobi. MATLAB and SciPy are usually based on high-level languages, which can limit performance when computational problems are required. GlobalOptimus is an algorithmic optimization tool written in C++ that combines high performance and ease of parameterization for solving optimization problems in many domains. While other optimization platforms exist, GlobalOptimus stands out for its openness and customizability.GlobalOptimus differs from other optimization tools in that it offers full support for using various optimization algorithms, such as genetic algorithms, differential evolution, and others.  Unlike tools like MATLAB or SciPy, GlobalOptimus allows users to work with an environment that is fully customizable and integrate software from other platforms without restrictions. The fact that it is open source enhances its modifiability.  Closing Gurobi which focuses on mathematical optimization requires a commercial license, while GlobalOptimus is open source.
 </p>
 
 ##  Software 
