### 2025-02-19 itsoulos/GlobalOptimus#6 @PaulDebus
Very cool. I sadly cannot test this since my computer runs out of RAM and has to kill the application after testing some of the optimization functions.

I dont know if those algorithms just need more RAM (I have 16GB of which around 2GB are available) or if this is a memory leak.

@nicoguaro what do you think, does this check the "automated tests" checkbox in the review? I think we are getting close
