+++ runrbf.sh
@@ -1,137 +1,0 @@
-#!/bin/bash
-DATAPATH=~/Desktop/ERGASIES/FeatureConstruction2/datasets/tenfolding/
-NODES=10
-
-#Available optimization methods: GradientDescent,Adam,Bfgs,Lbfgs,NelderMead,Genetic,Multistart,iPso,NeuralMinimizer,DifferentialEvolution, ParallelDe, Simman
-METHOD=Genetic
-#Available local search methods: bfgs, lbfgs, gradient, nelderMead, adam
-LOCALSEARCH=bfgs
-#Available samplers: uniform, mlp, rbf, maxwell, triangular, kmeans
-SAMPLER=uniform
-#Available stopping rules: maxiters, doublebox, similarity
-TERMINATION=doublebox
-
-if [ $METHOD = "Bfgs" ]
-then
-
-#bfgs_iters: the maximum number of allowed iterations
-
-	METHODPARAMS="--bfgs_iters=200"
-elif [ $METHOD = "NelderMead" ]
-then
-
-#nm_population: number of items in simplex
-#nm_alpha: the alpha parameter
-#nm_gamma: the gamma parameter
-#nm_rho:   the rho parameter
-#nm_sigma: the sigma parameter
-#nm_maxiters: maximum number of iterations allowed
-
-	METHODPARAMS="--nm_population=100 --nm_alpha=1.0 --nm_gamma=2.0 --nm_rho=0.5 --nm_sigma=0.5 --nm_maxiters=100"
-elif [ $METHOD = "Adam" ]
-then
-
-#adam_rate: adam learning rate
-#adam_linesearch: linesearch method used. Possible values (none|golden|fibonacci|armijo)
-#adam_maxiters: maximum number of iterations allowed
-#adam_b1: the b1 parameter
-#adam_b2: the b2 parameter
-	METHODPARAMS="--adam_rate=0.001 --adam_linesearch=armijo --adam_maxiters=10000 --adam_b1=0.9 --adam_b2=0.999"
-elif [ $METHOD = "GradientDescent" ]
-then
-
-#gd_maxiters: maximum number of allowed iterations
-#gd_linesearch: the linesearch method used (none|golden|fibonacci|armijo)
-#gd_epsilon: the epsilon value used in comparisons
-#gd_rate: the line search rate
-
-	METHODPARAMS="--gd_maxiters=200 --gd_linesearch=armijo --gd_epsilon=0.000001 --gd_rate=0.001"
-elif [ $METHOD = "Genetic" ]
-then
-
-#gen_lrate:		local search rate
-#gen_srate:		crossover rate
-#gen_mrate:		mutation rate
-#gen_tsize: 		tournament size
-#gen_selection:		selection scheme (tournament|roulette)
-#gen_crossover:		crossover scheme (double,uniform,onepoint)
-#gen_mutation:		mutation  scheme (double|random) 
-#gen_termination:	termination method (maxiters|doublebox|similarity) 
-#gen_count:		number of chromosomes
-#gen_maxiters:		maximum number of generations
-
-	METHODPARAMS="--gen_lrate=0.01 --gen_srate=0.1 --gen_mrate=0.05 --gen_tsize=8 --gen_selection=tournament --gen_crossover=double --gen_mutation=double --gen_termination=$TERMINATION --gen_count=200 --opt_sampler=$SAMPLER --opt_localsearch=$LOCALSEARCH --gen_maxiters=200 --opt_debug=yes"
-elif [ $METHOD = "Multistart" ]
-then
-
-#ms_samples: number of multistart samples
-#ms_maxiters: maximum number of iterations
[…]
