### 2025-02-20 bcf11e9
add functional rbf in DataFitting
