### 2025-02-20 ba267d1
bound values in frbf
