### 2025-02-23 79b4365
add parallel methods
