+++ METHODS/saop.cpp
@@ -1,0 +1,298 @@
+#include <METHODS/saop.h>
+
+SAOP::SAOP()
+{
+    before = std::chrono::system_clock::now();
+    addParam(Parameter("prop", 0, 0, 1, "0 or 1"));
+    addParam(Parameter("population", "100", "The number of elements."));
+    addParam(Parameter("maxiters", 100, 10, 10000, "Maximum iters"));
+    addParam(Parameter("localsearchRate", 0.01, 0.00, 1.0, "Localsearch rate"));
+    addParam(Parameter("subPopulation", "10", "The number of subelements."));
+    addParam(Parameter("propagationNumber", 0, 0, 50, "Number of positions to propagate."));
+    QStringList propagationMethod;
+    propagationMethod<<"1to1"<<"1toN"<<"Nto1"<<"NtoN";
+    addParam(Parameter("propagationMethod", propagationMethod[0],
+                       propagationMethod,
+                       "Propagation strategy: 1to1, 1toN, Nto1, NtoN"));
+    addParam(Parameter("subPopEnable", "1", "Number of subpopulations required to terminate."));
+    addParam(Parameter("propagationRate", "10", "Number of positions to propagate."));
+}
+
+void SAOP::init()
+{
+    prop = getParam("prop").getValue().toInt();
+    subPopulation = getParam("subPopulation").getValue().toInt();
+    dimension = myProblem->getDimension();
+    population = getParam("population").getValue().toInt();
+    maxIterations = getParam("maxiters").getValue().toInt();
+    localSearchRate = getParam("localsearchRate").getValue().toDouble();
+    propagationNumber = getParam("propagationNumber").getValue().toInt();
+    propagationMethod = getParam("propagationMethod").getValue();
+    subPopEnable = getParam("subPopEnable").getValue().toInt();
+    propagationRate = getParam("propagationRate").getValue().toInt();
+
+    printf("prop=%d\tpropagationNumber=%d\tpropagationRate=%d\tpropagationMethod=%s\n",prop
+           ,propagationNumber
+           ,propagationRate
+           ,propagationMethod.toStdString().c_str());
+
+    iteration = 0;
+    bestF2x.resize(subPopulation);
+    bestF2xOld.resize(subPopulation);
+    bSamply.resize(dimension);
+    similarityCurrentCountBest.resize(subPopulation);
+    for (int k = 0; k < subPopulation; k++)
+    {
+        bestF2x.at(k) = 1e+100;
+        bestF2xOld.at(k) = 1e+100;
+        similarityCurrentCountBest.at(k) = 0;
+    }
+    bF2x = 1e+100;
+    smellIntensity.resize(population);
+    evaporationRate = 0.01;
+    sampleFromProblem(population, samples, fitnessArray);
+    bestSamply.resize(subPopulation);
+    sub_size = population / subPopulation;
+    for (int k = 0; k < subPopulation; k++)
+    {
+        int start = k * sub_size;
+        int end = start + sub_size;
+        for (int i = start; i < end; i++)
+        {
+            Data x = samples.at(i);
+            double y = fitnessArray.at(i);
+            if (i == 0 || y < bestF2x.at(k))
+            {
+                bestF2x.at(k) = y;
+                bestSamply.at(k) = x;
+            }
+            if (i == 0 || y < bF2x)
+            {
+                bF2x = y;
+                bSamply = x;
+            }
+            smellIntensity.at(i) = calculateSmellIntensity(fitnessArray.at(i));
+        }
+    }
+}
+
+double SAOP::randomDouble(double min, double max)
+{
+    return min + (max - min) * (rand() / (double)RAND_MAX);
+}
+
[…]
