+++ METHODS/acop.cpp
@@ -1,0 +1,330 @@
+#include <METHODS/acop.h>
+
+ACOP::ACOP()
+{
+    before = std::chrono::system_clock::now();
+    addParam(Parameter("population", "100", "The number of elements."));
+    addParam(Parameter("maxiters", 100, 10, 10000, "Maximum iters"));
+    addParam(Parameter("localsearchRate", 0.01, 0.00, 1.0, "Localsearch rate"));
+    addParam(Parameter("subPopulation", "4", "Number of subpopulations"));
+    addParam(Parameter("subPopEnable", "1", "Number of subpopulations required to terminate."));
+    addParam(Parameter("prop", 0, 0, 1, "0 or 1"));
+    addParam(Parameter("propagationNumber", 0, 0, 50, "Number of positions to propagate."));;
+    QStringList propagationMethod;
+    propagationMethod<<"1to1"<<"1toN"<<"Nto1"<<"NtoN";
+    addParam(Parameter("propagationMethod", propagationMethod[0],
+                       propagationMethod,
+                       "Propagation strategy: 1to1, 1toN, Nto1, NtoN"));
+    addParam(Parameter("propagationRate", "10", "Number of positions to propagate."));
+}
+
+void ACOP::init()
+{
+    dimension = myProblem->getDimension();
+    population = getParam("population").getValue().toInt();
+    maxIterations = getParam("maxiters").getValue().toInt();
+    localSearchRate = getParam("localsearchRate").getValue().toDouble();
+    subPopulation = getParam("subPopulation").getValue().toInt();
+    subPopEnable = getParam("subPopEnable").getValue().toInt();
+    prop = getParam("prop").getValue().toInt();
+    propagationNumber = getParam("propagationNumber").getValue().toInt();
+    propagationMethod = getParam("propagationMethod").getValue();
+    propagationRate = getParam("propagationRate").getValue().toInt();
+
+    printf("prop=%d\tpropagationNumber=%d\tpropagationRate=%d\tpropagationMethod=%s\n",prop
+           ,propagationNumber
+           ,propagationRate
+           ,propagationMethod.toStdString().c_str());
+    //printf("propagationMethod=%s\n",propagationMethod.toUtf8().data());
+
+
+
+    iteration = 0;
+    alpha = 1.0;
+    beta = 2.0;
+    evaporationRate = 0.5;
+    Q = 100.0;
+    bestF2xAll = 1e+100;
+
+    sub_size = population / subPopulation;
+    propagationNumber = min(propagationNumber, sub_size);
+
+    antPositions.resize(population, vector<double>(dimension));
+    fitnessArray.resize(population);
+    pheromones.resize(population, vector<double>(dimension, 1.0));
+    bestF2x.resize(subPopulation, 1e+100);
+    bestSamply.resize(subPopulation, vector<double>(dimension));
+    bestF2xOld.resize(subPopulation, 1e+100);
+    similarityCurrentCountBest.resize(subPopulation, 0);
+    // bestAllSamply.resize(dimension);
+    sampleFromProblem(population, antPositions, fitnessArray);
+
+#pragma omp parallel for num_threads(subPopulation)
+    for (int k = 0; k < subPopulation; k++)
+    {
+        int start = k * sub_size;
+        int end = min(start + sub_size, population); // Προστασία out-of-bounds
+
+        for (int i = start; i < end; i++)
+        {
+            double y = fitnessArray[i];
+            if (y < bestF2x[k])
+            {
+                bestF2x[k] = y;
+                bestSamply[k] = antPositions[i];
+            }
+        }
+    }
+}
+
+double ACOP::randomDouble(double min, double max)
+{
+    return min + (max - min) * (rand() / (double)RAND_MAX);
+}
+
+void ACOP::moveAnts()
+{
[…]
