### 2025-02-26 itsoulos/GlobalOptimus#1 @PaulDebus
I just made a pull request with some minor changes to the paper. Please check them out. I also have some more ideas to improve the paper:
- You should take another look and check if you are happy with all the claims. You make some very strong claims about Scipy and Matlab that not everybody would agree on. For me personally, a Python environment is very easy to use and customizable, a lot easier than having to build a complex application relying on Qt
- Figure 1 needs an update. There is a typo in the figure (`terminate` instead of `terminated`) and the resolution is very low. You could either make the figure smaller, provide a higher resolution, or leave it out altogether - the table above already shows the same content. You could even extend the table to show the return type (or something) of those functions, so it becomes clearer - I think there is some horizontal space left.
- I think an interesting feature of your tool is to allow users to code their own problems. You could provide the documentation for this similar how you provide the documentation for coding new optimization methods.

Once we get the remaining issues listed in the first comment fixed, we can proceed. For the testing, I am waiting for feedback of @nicoguaro
