### 2025-02-26 #7584 @PaulDebus
I think my review has moved on to the smaller things and improvements of the actual paper.

The topic of automated testing is still not 100% done but we are getting there
