### 2025-02-27 itsoulos/GlobalOptimus#6 @itsoulos
@PaulDebus   I have just installed the latest GlobalOptimus on a fedora linux running on a machine with 8GB of ram. The testing procedure completed without any issue. The command line for GlobalOptimus that I have used is:
./GlobalOptimus --opt_problem=bf1 --opt_enabletesting=yes

Could you please share your command in order to check if there is any memory issue?
