+++ README.md
@@ -100,32 +100,32 @@ From this dialog, the user can select the global optimization method from a drop
 
 ## Available Global Optimization methods
 The implemented Global Optimization methods are:
-1. The Differential Evolution denoted as DifferentialEvolution.
-2. The Parallel Differential Evolution method (utilizing the OpenMP library), denoted as ParallelDe.
-3. A genetic algorithm, denoted as Genetic in the package.
-4. An improved version of the Particle Swarm Optimizer, denoted as iPSO
-5. A simple Multistart method, which initiates many local optimization procedures to discover the global optimum of a function.
-6. The NeuralMinimizer method, which constructs an estimated function of the objective problem using Neural Networks.
-7. A parallel Particle Swarm Optimizer (using the OpenMP library), denoted as ParallelPso.
-8. A Simulated Annealing optimizer, denoted as Simman.
-9. The optimal foraging algorithm (OFA), donated as Ofa.
-10. Bio-inspired metaheuristic algorithm Giant Armadillo Optimization (GAO), donated as armadillo1.
-11. The Gray Wolf Optimizer (GWO), donated as Gwo.
+1. The Differential Evolution denoted as **DifferentialEvolution**.
+2. The Parallel Differential Evolution method (utilizing the OpenMP library), denoted as **ParallelDe**.
+3. A genetic algorithm, denoted as **Genetic** in the package.
+4. An improved version of the Particle Swarm Optimizer, denoted as **iPSO**.
+5. A simple **Multistart** method, which initiates many local optimization procedures to discover the global optimum of a function.
+6. The **NeuralMinimizer** method, which constructs an estimated function of the objective problem using Neural Networks. 
+7. A parallel Particle Swarm Optimizer (using the OpenMP library), denoted as **ParallelPso**.
+8. A Simulated Annealing optimizer, denoted as **Simman**.
+9. The optimal foraging algorithm (OFA), donated as **Ofa**.
+10. Bio-inspired metaheuristic algorithm Giant Armadillo Optimization (GAO), denoted as **armadillo1**.
+11. The Gray Wolf Optimizer (GWO), donated as **Gwo**.
 
 ## Available Local Optimization methods
 All global optimization methods can be enhanced by applying a local minimization method after they are terminated. The parameter used to determine the used local optimization procedure is the --opt_localsearch parameter. The implemented local optimization methods are the following:
-1. The bfgs method. The Broyden–Fletcher–Goldfarb–Shanno (BFGS) algorithm was implemented using a variant of Powell.
-2. The limited memory BFGS method (lbfgs).
-3. The Gradient descent method. This method is denoted as gradient.
-4. The Nelder Mead method (denoted as nelderMead).
-5. The Adam optimizer (denoted as adam).
+1. The **bfgs** method. The Broyden–Fletcher–Goldfarb–Shanno (BFGS) algorithm was implemented using a variant of Powell.
+2. The limited memory BFGS method (**lbfgs**).
+3. The Gradient descent method. This method is denoted as **gradient**.
+4. The Nelder Mead method (denoted as **nelderMead**).
+5. The Adam optimizer (denoted as **adam**).
 
 ## Problem example
 Consider the Rastrigin problem, defined as: 
  $f(x)=x_{1}^{2}+x_{2}^{2}-\cos(18x_1)-\cos(18x_2)$. This function is implemented as rastriginproblem.cpp in the PROBLEMS subdirectory of the distribution.  The RastriginProblem class contains the following main methods:
- 1. The constructor method RastriginProblem. In this function dimension of the objective function should be 
+ 1. The constructor method `RastriginProblem`. In this function dimension of the objective function should be 
[…]
