### 2025-02-27 itsoulos/GlobalOptimus#1 @itsoulos
@PaulDebus  the feature of creating user problems is fully described in README.md. As you known there is a limitation in the length of PAPER.md and therefore we can not add this description there.
