### 2025-02-27 itsoulos/GlobalOptimus#6 @PaulDebus
I ran the tests again after getting the latest updates and rebuilding the binary and now it worked. But I had a bit more RAM available, maybe that helped. I also tested with the rastrigin problem, which last time did crash.

```
./GlobalOptimus  --opt_problem=rastrigin --opt_enabletesting=yes
```

Maybe it was an issue of me not having much free RAM at that time.
