### 2025-02-27 itsoulos/GlobalOptimus#1 @PaulDebus
Yes, the description is in the Readme and very helpful there. I was just thinking that instead of the figure, you could add the description how to define a user problem to the paper, as this is an interesting feature the users should be aware of.

Just a suggestion, if you want to keep the figure, this is absolutely fine by me - try to make it in a better resolution so it looks nice
