### 2025-02-27 itsoulos/GlobalOptimus#6 @PaulDebus
I tested again and the tests are running now. So no problem there.

I just want to discuss with nicoguaro if we can check the "automated testing" requirement for the paper
