+++ doc/parallelSAO/SUBMIT_SAOP/saop.tex
@@ -1,0 +1,2412 @@
+%% LyX 2.4.1 created this file.  For more info, see https://www.lyx.org/.
+%% Do not edit unless you really know what you are doing.
+\documentclass[journal,article,submit,pdftex,moreauthors]{Definitions/mdpi}
+\usepackage[utf8]{inputenc}
+\usepackage{cprotect}
+\usepackage{float}
+\usepackage{url}
+\usepackage{varwidth}
+\usepackage{graphicx}
+
+\makeatletter
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% LyX specific LaTeX commands.
+
+\Title{Parallel SAO: Collaborative Subpopulations for Accelerated Convergence}
+
+\TitleCitation{Parallel SAO: Collaborative Subpopulations for Accelerated Convergence}
+
+\Author{Glykeria Kyrou$^{1}$, Ioannis G. Tsoulos$^{2,*}$ , Anna Maria Gianni$^{3}$
+and Vasileios Charilogis$^{4}$ }
+
+\AuthorNames{Glykeria Kyrou, Ioannis G. Tsoulos, A.M. Gianni and Vasileios Charilogis}
+
+\AuthorCitation{Kyrou, G.; Tsoulos, I.G.; Gianni A.M; Charilogis, V. }
+
+
+\address{$^{1}$\quad{}Department of Informatics and Telecommunications,
+University of Ioannina, 47150 Kostaki Artas, Greece; g.kyrou@uoi.gr\\
+$^{2}$\quad{}Department of Informatics and Telecommunications, University
+of Ioannina, 47150 Kostaki Artas, Greece;itsoulos@uoi.gr\\
+$^{3}$\quad{}Department of Informatics and Telecommunications, University
+of Ioannina, 47150 Kostaki Artas, Greece; am.gianni@uoi.gr\\
+$^{4}$\quad{}Department of Informatics and Telecommunications, University
+of Ioannina, 47150 Kostaki Artas, Greece; v.charilog@uoi.gr}
+
+
+\corres{Correspondence: itsoulos@uoi.gr}
+
+
+\abstract{In the dynamically evolving field of collective computational optimization,
+modern approaches increasingly incorporate bio-inspired techniques,
+such as Smell Agent Optimization (SAO), to address complex, high-dimensional
+problems inherent to contemporary scientific and industrial applications.
+While these methods are distinguished by their dynamic convergence
+and heuristic ability to explore vast solution spaces, their growing
+computational complexity hinders their application in real-world,
+large-scale scenarios where simultaneous speed and precision are critical.
+To overcome this challenge, the present research advances a pioneering
+parallel implementation of SAO, which transcends simple workload distribution
+by integrating dynamic collaboration mechanisms and intelligent information
+dispersal among autonomous subpopulations. Concurrently, the method
+is enriched with innovative rules for exchanging optimal solutions
+between subpopulations. These rules not only prevent premature convergence
+to local minima but also establish a continuous flow of information
+that accelerates the global exploration of the solution space. Experimental
+validation of the proposed method demonstrated that, through optimized
+parameterization of the diffusion mechanisms, SAO's efficiency can
+exceed 50\%, achieving simultaneous reductions in both the number
+of objective function evaluations and total execution time. This outcome
+holds particular significance in high-dimensional problems, where
+balancing computational cost and accuracy is a decisive factor. These
+findings not only underscore the potential of parallel SAO to deliver
+sustainable solutions to real-world challenges but also open new horizons
+in the theory and practice of collective optimization. The implications
+extend to domains such as large-scale data analysis, autonomous systems,
[…]
