### 2025-03-26 itsoulos/GlobalOptimus#1 @PaulDebus
Looks good. Maybe you can reference the new UserProblem diagram in the first paragraph of the "Usage of the Software" section.

Also you could reduce the size of the figures 1 and 3 to save a bit of space. The required markdown is given in the documentation: https://joss.readthedocs.io/en/latest/paper.html#figures-and-images

But other than that, it looks good now.
