### 2025-03-26 #7584 @PaulDebus
My review is nearly done. The only thing missing is the topic of automated testing, where I am waiting to read what @nicoguaro has to say. After that, I can probably complete my review.
