+++ paper/paper.md
@@ -51,6 +51,9 @@ Finally, Gurobi is an extremely powerful tool for mathematical optimization, but
 <p style="text-align: justify;">
 The software is compiled using the QT library, available from  https://qt.io. The researcher should code their objective function and a number of other mandatory functions in the C++ programming language, specifying the dimension of the objective function and the bounds of the function domain. Subsequently, they can select a global optimization method to apply to the problem from a wide range of available methods.</p>
 
+In addition, the user can create his own objective function by implementing the methods of category `UserProblem`,
+the methods of which are presented in the following diagram.
+![The diagram of the user defined problem UserProblem \label{FIG:userproblem}](userproblem.png)
 ## Implemented global optimization methods
 <p style="text-align: justify?">
  In the proposed software, each implemented global optimization method has a set of parameters that can determine the overall optimization path and the effectiveness of the method. Some of the global optimization methods are as follows: </p>
@@ -104,10 +107,6 @@ A flowchart of any used optimization process is outlined in \autoref{FIG:optimiz
 An example run for the Rastrigin function and the Genetic algorithm is outlined in \autoref{FIG:rastrigin}.
 
 ![An example run for the Rastrigin function \label{FIG:rastrigin}](rastrigin.png)
-In addition, the user can create his own objective function by implementing the methods of category `UserProblem`,
-the methods of which are presented in the following diagram.
-![The diagram of the user defined problem UserProblem \label{FIG:userproblem}](userproblem.png)
-
 
 #  Conclusions
 <p style="text-align: justify;">
