### 2025-03-27 itsoulos/GlobalOptimus#1 @itsoulos
We have updated the figures and make the corrections in paper.md file
