### 2025-03-27 itsoulos/GlobalOptimus#1 @PaulDebus
i made a pull request for image formatting - but i am not sure it works. you can take it if you want.

but i am only waiting on nico and then we can complete the review
