+++ doc/NNC_GA/SUBMIT_NNC_GA/nnc_ga.tex
@@ -1,4 +1,4 @@
-%% LyX 2.3.7 created this file.  For more info, see http://www.lyx.org/.
+%% LyX 2.4.1 created this file.  For more info, see https://www.lyx.org/.
 %% Do not edit unless you really know what you are doing.
 \documentclass[journal,article,submit,pdftex,moreauthors]{Definitions/mdpi}
 \usepackage[utf8]{inputenc}
@@ -35,32 +35,36 @@ University of Western Macedonia, 50100 Kozani, Greece; tsalikakis@gmail.com}
 \corres{Correspondence: itsoulos@uoi.gr}
 
 
-\abstract{A machine learning technology that has been used in a wide range
-of everyday problems is the recently proposed technique of constructing
-artificial neural networks with the assistance of Grammatical Evolution.
-This technique was utilized in problems from the fields of physics,
-chemistry, medicine, etc., in which it had excellent performance compared
-to other machine learning models. A key feature of these machine learning
-models is that they can both construct the structure of an artificial
-neural network efficiently and identify optimal values of the parameters
-of this network at the same time. However, in many cases the optimization
-process can become trapped in local minima of the error function,
-resulting in the construction of the artificial neural network not
-being able to be completed satisfactorily, which will lead to poor
-results in terms of test error. In this work, the combination of the
-artificial neural network construction method with an adapted genetic
-algorithm is proposed, which preserves the structure of the neural
-network that has been constructed and, in addition, appropriately
-adapts the parameters of the artificial neural network in order to
-avoid overfitting problems. The proposed method was applied on a wide
-series of datasets from real world problems that include classification
-and regression problems and it was compared against traditional machine
-learning models and the results are reported.}
+\abstract{This study presents a novel hybrid approach combining grammatical
+evolution with constrained genetic algorithms to overcome key limitations
+in automated neural network design. The proposed method addresses
+two critical challenges: the tendency of grammatical evolution to
+converge to suboptimal architectures due to local optima, and the
+common overfitting problems in evolved networks. Our solution employs
+grammatical evolution for initial architecture generation while implementing
+a specialized genetic algorithm that simultaneously optimizes network
+parameters within dynamically adjusted bounds. The genetic component
+incorporates innovative penalty mechanisms in its fitness function
+to control neuron activation patterns and prevent overfitting. Comprehensive
+testing across 53 diverse datasets shows our method achieves superior
+performance compared to traditional optimization techniques, with
+an average classification error of 21.18\% versus 36.45\% for ADAM,
+while maintaining better generalization capabilities. The constrained
+optimization approach proves particularly effective in preventing
+premature convergence, and the penalty system successfully mitigates
+overfitting even in complex, high-dimensional problems. Statistical
+validation confirms these improvements are significant (p \textless{}
+1.1e-08) and consistent across multiple domains including medical
+diagnosis, financial prediction, and physical system modeling. This
+work provides a robust framework for automated neural network construction
+that balances architectural innovation with parameter optimization
+while addressing fundamental challenges in evolutionary machine learning.}
 
 
 \keyword{Grammatical Evolution; Genetic Programming; Neural networks; Local
 Optimization}
 
+\newcommand*\LyXZeroWidthSpace{\hspace{0pt}}
 \DeclareTextSymbolDefault{\textquotedbl}{T1}
 %% Because html converters don't know tabularnewline
 \providecommand{\tabularnewline}{\\}
[…]
