### 2025-04-08 #7584 @nicoguaro
@diehlpk, I have been making progress in the review (although very slowly). Currently, it worries me (besides the automatic testing) that I can't find any documentation of the software in the repo. There is a folder called ``doc``, but I do not know if it compiles or generates some document of any type since it does not have a README.

The authors mention in the paper that optimization suites based on high-level languages might suffer from performance issues, but I do not see a tutorial on how to implement your own problem. Based on what the main README has, I assume that one can program a class for its own problem and recompile the whole program to achieve it. But from the user's perspective, a step-by-step description is needed. Also, from the usability perspective, it would be nice to change the optimization problem without the need to recompile the whole thing.
