### 2025-04-08 #7584 @itsoulos
@nicoguaro  we have a detailed readme file in the repo. The doc file contains papers that we have published in the recent years using the GlobalOptimus. Also, we have detailed information on the paper on how to implement your own problem and use it in the program. Finally, it is necessary to recompile the software in order to include your own problem, otherwise we have to use shared files and this will produce additional difficulties on various operating systems. For now the usage of the QT environment is enough to recompile and use the software from the command line using a limited number of steps
