### 2025-04-08 #7584 @itsoulos
Also, regarding automate testing we have added the option opt_enabletesting in the main program following your advice about it. This option could be used to test a series of optimization methods on a selected function. Hence, I do not know what else do you mean by the term "automate testing".
