### 2025-04-08 itsoulos/GlobalOptimus#6 @diehlpk
> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I think there is no requirement to have some CI, however, having a script or using ctest to run the tests should be there. If there is some documentation on how to run the tests using a script, like `run_test.sh` should be fine. However, the user should not need run all test by hand or write own tests.
