### 2025-04-08 itsoulos/GlobalOptimus#6 @itsoulos
@diehlpk  we have already a script to run the tests without the need for parameters. Could you please check again the README?
