### 2025-04-08 #7584 @diehlpk
@itsoulos To test the software, I have to run

`./GlobalOptimus  --opt_problem=test2n --opt_method=Genetic --opt_debug=no --opt_iters=10 --opt_dimension=5 --gen_lrate=0.05`

For all available methods on my own or is there some script to run the tests?
