### 2025-04-08 #7584 @diehlpk
> [@diehlpk](https://github.com/diehlpk) it is easy for us to use doxygen and I think that actually it is the only issue here. We have documentation regarding user problem in the README and of course we have added automating test using a new parameter. Hence, we are going to add doxygen output from the classes.

Yes, please add doxygen comments and instructions on how to generate the doc. That should resolve the documentation issue.
