### 2025-04-08 #7584 @diehlpk
> [@diehlpk](https://github.com/diehlpk) No. There is the script runfunmin.sh that the user can use to test methods and functions. Could please locate it in the README?

No that does not count for automated testing. With automated testing the user should be able to run some tests to check if the code was correctly installed. The user should not need to write his own tests since the user does not know the software. Can you write a script which runs a few tests and so the user knows that the code is installed correctly.
