### 2025-04-08 #7584 @itsoulos
What do you mean correctly installed? The software is just an executable with many command line options. We have added the parameter opt_enabletesting in the command line options in order to apply  automatically a series  of global optimization methods on a selected test function. What else should we do?
