### 2025-04-08 #7584 @diehlpk
Just provide a script for example `run_test.sh` to run these tests.

A normal process is

1. Configure software, e.g. cmake or any other tool
2. Compile software, e.g. make or any other tool
3. Install software, e.g. make install
4. Run tests, e.g. ctest or any other tool

I think having some script to run the tests would be helpful. You could just run the executable with the  parameter opt_enabletesting in the script.
