### 2025-04-08 #7584 @itsoulos
@diehlpk  Hence, you need a script that
1) executes the qmake to create the Makefile
2) executes the make command
3) execute the testing on a selected test function using the opt_enabletesting
Is this correct?
