### 2025-04-08 itsoulos/GlobalOptimus#6 @diehlpk
@itsoulos Please provide an example on how I would run the tests, I am confused.
