### 2025-04-08 itsoulos/GlobalOptimus#6 @itsoulos
@diehlpk  you could use the runfunmin.sh as
./runfunmin.sh camel

Also, you may edit the line GLOBALPARAMS in the runfunmin.sh to add the option --opt_enabletesting=yes in this line or
 you may run
./GlobalOptimus  --opt_problem=camel --opt_enabletesting=yes
