### 2025-04-08 #7584 @diehlpk
My issue is that if I read https://github.com/itsoulos/GlobalOptimus?tab=readme-ov-file#testing there is no mention of a script and I assume that I have to call

`./GlobalOptimus  --opt_problem=test2n --opt_method=Genetic --opt_debug=no --opt_iters=10 --opt_dimension=5 --gen_lrate=0.05`

and provide all options for all methods. If there is a script that I just can, please mention it in the testing section.
