### 2025-04-08 #7584 @diehlpk
> [@diehlpk](https://github.com/diehlpk) hence to summarize it
>
>     1. we add the documentation using doxygen
>
>     2. we add a script with these 3 steps
>
>
> Are these enough for the paper?

@itsoulos Yes, adding documentation is required, at least of the core API and please add to the documentation on how to generate it.
