### 2025-04-08 #7584 @itsoulos
@diehlpk  yes, but we have added the section  "Comparing with known global optima" just below for the new added option. Is there a need to merge these sections?
