### 2025-04-08 itsoulos/GlobalOptimus#6 @diehlpk
@itsoulos Ok, no things are more clear.

I would prefer a script which calls

`./GlobalOptimus --opt_problem=problem1 --opt_enabletesting=yes`
`./GlobalOptimus --opt_problem=problem2 --opt_enabletesting=yes`

For all of the provided problems and I think we are fine with testing and mention this script in the testing section.
