### 2025-04-08 itsoulos/GlobalOptimus#6 @itsoulos
@diehlpk  we will added these script also and a clear mention will be added in the README
