### 2025-04-08 itsoulos/GlobalOptimus#6 @diehlpk
@nicoguaro and @PaulDebus Please let me know if you agree with adding a script to run the tests and mention that in the testing sections. That should remove the automated test issue.
