### 2025-04-09 34e3a41
add comments in the classes of the package
