+++ GE/integeranneal.h
@@ -3,6 +3,10 @@
 # include <OPTIMUS/problem.h>
 # include <OPTIMUS/optimizer.h>
 # include <GE/program.h>
+/**
+ * @brief The IntegerAnneal class implements
+ * the simulated annealing algorithm for Grammatical Evolution problems.
+ */
 class IntegerAnneal
 {
 private:
@@ -16,12 +20,40 @@ private:
     double T0;
     int k;
 public:
+    /**
+     * @brief IntegerAnneal The constructor of the class for Program pr.
+     * @param pr
+     */
     IntegerAnneal(Program *pr);
+    /**
+     * @brief setNeps Sets the number of samples
+     * @param n
+     */
     void    setNeps(int n);
+    /**
+     * @brief setT0 Sets the initial temperature.
+     * @param t
+     */
     void    setT0(double t);
-    void setPoint(vector<int> &g,double &y);
-    void getPoint(vector<int> &g,double &y);
-    void Solve();
+    /**
+     * @brief setPoint Sets the initial point of the algorithm.
+     * @param g
+     * @param y
+     */
+    void    setPoint(vector<int> &g,double &y);
+    /**
+     * @brief getPoint Returns the final point located by the algorithm.
+     * @param g
+     * @param y
+     */
+    void    getPoint(vector<int> &g,double &y);
+    /**
+     * @brief Solve Executes the main loop of the Simulated Annealing.
+     */
+    void    Solve();
+    /**
+     * @brief updateTemp Updates the temperature.
+     */
     void    updateTemp();
     ~IntegerAnneal();
 };
