+++ METHODS/ipso.h
@@ -3,6 +3,10 @@
 
 # include <OPTIMUS/collection.h>
 # include <OPTIMUS/optimizer.h>
+
+/**
+ * @brief The iPso class implements the Improved PSO optimizer.
+ */
 class iPso : public Optimizer
 {
 
@@ -26,19 +30,55 @@ private:
     double RC;
     int localSearchCount;
 
+    /**
+     * @brief fitness
+     * @param x
+     * @return the fitness value of particle x.
+     */
     double fitness(Data &x);
+    /**
+     * @brief checkGradientCriterion
+     * @param x
+     * @return Check if gradient criterion holds for particle x.
+     */
     bool checkGradientCriterion(Data &x);
-
+    /**
+     * @brief calcFitnessArray Calculates the fitness values for all particles.
+     */
     void calcFitnessArray();
+    /**
+     * @brief updateBest Updates the best particle.
+     */
     void updateBest();
+    /**
+     * @brief updateCenter Updates the center particle (for center pso).
+     */
     void updateCenter();
-
 public:
+    /**
+     * @brief iPso The constructor of the class.
+     */
     iPso();
+    /**
+     * @brief init Initializes the particles and velocities.
+     */
     virtual void init();
+    /**
+     * @brief step Performs a step of the PSO.
+     */
     virtual void step();
+    /**
+     * @brief terminated
+     * @return true when PSO should be terminated.
+     */
     virtual bool terminated();
+    /**
+     * @brief showDebug Displays debug information.
+     */
     virtual void showDebug();
+    /**
+     * @brief done It is called when PSO finishes.
+     */
     virtual void done();
     virtual ~iPso();
 };
