### 2025-04-09 8977763
add doxygen manual
