+++ MANUAL/docs/latex/classMlpProblem.tex
@@ -1,0 +1,418 @@
+\hypertarget{classMlpProblem}{}\doxysection{Mlp\+Problem Class Reference}
+\label{classMlpProblem}\index{MlpProblem@{MlpProblem}}
+
+
+The \mbox{\hyperlink{classMlpProblem}{Mlp\+Problem}} class implements the neural network (MLP) model.  
+
+
+
+
+{\ttfamily \#include $<$mlpproblem.\+h$>$}
+
+
+
+Inheritance diagram for Mlp\+Problem\+:
+\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[width=211pt]{classMlpProblem__inherit__graph}
+\end{center}
+\end{figure}
+
+
+Collaboration diagram for Mlp\+Problem\+:
+\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[width=350pt]{classMlpProblem__coll__graph}
+\end{center}
+\end{figure}
+\doxysubsection*{Public Member Functions}
+\begin{DoxyCompactItemize}
+\item 
+\mbox{\Hypertarget{classMlpProblem_a1dee677489db08feff6d547f1bf6d3cc}\label{classMlpProblem_a1dee677489db08feff6d547f1bf6d3cc}} 
+{\bfseries Mlp\+Problem} ()
+\begin{DoxyCompactList}\small\item\em \mbox{\hyperlink{classMlpProblem}{Mlp\+Problem}} The main constructor of MLP. \end{DoxyCompactList}\item 
+void \mbox{\hyperlink{classMlpProblem_a8e49f08fa0685c93645fd83684bbecf5}{set\+Weights}} (Data \&w)
+\begin{DoxyCompactList}\small\item\em set\+Weights Sets the weights of MLP. \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{classMlpProblem_aaa6890783e47f1f99c92b664e7f60958}{funmin}} (Data \&x)
+\begin{DoxyCompactList}\small\item\em funmin \end{DoxyCompactList}\item 
+Data \mbox{\hyperlink{classMlpProblem_a65173bf310d7aa3b364746517f23afe5}{gradient}} (Data \&x)
+\begin{DoxyCompactList}\small\item\em gradient \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{classMlpProblem_a6aaaff4a29f3518842280dcebd5ade9f}{sig}} (double x)
+\begin{DoxyCompactList}\small\item\em sig \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{classMlpProblem_a4b5a4471b561b86adad7d30879f8dce5}{sigder}} (double x)
+\begin{DoxyCompactList}\small\item\em sigder \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{classMlpProblem_a73f28617d4a0e03e12e8bf42ae4a5442}{get\+Output}} (Data \&x)
+\begin{DoxyCompactList}\small\item\em get\+Output \end{DoxyCompactList}\item 
+Data \mbox{\hyperlink{classMlpProblem_a48a2248ff05d811440b1760fd57c19b2}{get\+Derivative}} (Data \&x)
+\begin{DoxyCompactList}\small\item\em get\+Derivative \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{classMlpProblem_a2b4b5f9b7503cd9d236d6af46a722da7}{get\+Derivative1}} (vector$<$ double $>$ xpoint, int pos)
[…]
