+++ MANUAL/docs/latex/classTriangularSampler.tex
@@ -1,0 +1,141 @@
+\hypertarget{classTriangularSampler}{}\doxysection{Triangular\+Sampler Class Reference}
+\label{classTriangularSampler}\index{TriangularSampler@{TriangularSampler}}
+
+
+The \mbox{\hyperlink{classTriangularSampler}{Triangular\+Sampler}} class samples point from the objective problem using triangular distribution.  
+
+
+
+
+{\ttfamily \#include $<$triangularsampler.\+h$>$}
+
+
+
+Inheritance diagram for Triangular\+Sampler\+:
+\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[width=175pt]{classTriangularSampler__inherit__graph}
+\end{center}
+\end{figure}
+
+
+Collaboration diagram for Triangular\+Sampler\+:
+\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[width=181pt]{classTriangularSampler__coll__graph}
+\end{center}
+\end{figure}
+\doxysubsection*{Public Member Functions}
+\begin{DoxyCompactItemize}
+\item 
+\mbox{\Hypertarget{classTriangularSampler_ad23349a0baf210972c3eefd593a26c22}\label{classTriangularSampler_ad23349a0baf210972c3eefd593a26c22}} 
+{\bfseries Triangular\+Sampler} (\mbox{\hyperlink{classProblem}{Problem}} $\ast$p)
+\item 
+void \mbox{\hyperlink{classTriangularSampler_a0d477fd8f7fa4ee8059241e4b2554c1d}{sample\+From\+Problem}} (int N, Matrix \&xsample, Data \&ysample)
+\begin{DoxyCompactList}\small\item\em sample\+From\+Problem, takes samples from the problem using some distribution. \end{DoxyCompactList}\item 
+void \mbox{\hyperlink{classTriangularSampler_a7823202e0c37b943f3f2c7986c40228e}{sample\+From\+Model}} (int \&N, Matrix \&xsample, Data \&ysample)
+\begin{DoxyCompactList}\small\item\em sample\+From\+Model, this method will sample from the sample model instead of the real problem. \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{classTriangularSampler_a22b94bd8f35b0f50264edadeae423e77}{eval}} (Data \&xpoint)
+\begin{DoxyCompactList}\small\item\em eval gets the value of model at point \end{DoxyCompactList}\item 
+void \mbox{\hyperlink{classTriangularSampler_acd5d903ad83f2d5bf90ad6d5b2093d72}{add\+Sample\+From\+Problem}} (Data \&x, double y)
+\begin{DoxyCompactList}\small\item\em add\+Sample\+From\+Problem adds a new sample from the problem. \end{DoxyCompactList}\end{DoxyCompactItemize}
+\doxysubsection*{Additional Inherited Members}
+
+
+\doxysubsection{Detailed Description}
+The \mbox{\hyperlink{classTriangularSampler}{Triangular\+Sampler}} class samples point from the objective problem using triangular distribution. 
+
+\doxysubsection{Member Function Documentation}
+\mbox{\Hypertarget{classTriangularSampler_acd5d903ad83f2d5bf90ad6d5b2093d72}\label{classTriangularSampler_acd5d903ad83f2d5bf90ad6d5b2093d72}} 
[…]
