+++ MANUAL/docs/latex/classMapper.tex
@@ -1,0 +1,20 @@
+\hypertarget{classMapper}{}\doxysection{Mapper Class Reference}
+\label{classMapper}\index{Mapper@{Mapper}}
+\doxysubsection*{Public Member Functions}
+\begin{DoxyCompactItemize}
+\item 
+\mbox{\Hypertarget{classMapper_a639409c1b280214479c017f47e22d737}\label{classMapper_a639409c1b280214479c017f47e22d737}} 
+{\bfseries Mapper} (int d)
+\item 
+\mbox{\Hypertarget{classMapper_a5116d056cd8a0274d4ab019e98ba6aec}\label{classMapper_a5116d056cd8a0274d4ab019e98ba6aec}} 
+bool {\bfseries set\+Expr} (vector$<$ string $>$ \&s)
+\item 
+\mbox{\Hypertarget{classMapper_ab6ac3a2a26b27cc155313152a90325f7}\label{classMapper_ab6ac3a2a26b27cc155313152a90325f7}} 
+int {\bfseries map} (Data \&x, Data \&x1)
+\item 
+\mbox{\Hypertarget{classMapper_aea29bdfb58b51d0e53551fd056339cea}\label{classMapper_aea29bdfb58b51d0e53551fd056339cea}} 
+bool {\bfseries map\+Dataset} (\mbox{\hyperlink{classDataset}{Dataset}} $\ast$original, \mbox{\hyperlink{classDataset}{Dataset}} $\ast$mapped\+Dataset)
+\item 
+\mbox{\Hypertarget{classMapper_a1b06305b08b93a7c6716bf653db98e1b}\label{classMapper_a1b06305b08b93a7c6716bf653db98e1b}} 
+bool {\bfseries map\+And\+Replace\+Dataset} (\mbox{\hyperlink{classDataset}{Dataset}} $\ast$original, \mbox{\hyperlink{classDataset}{Dataset}} $\ast$mapped\+Dataset)
+\end{DoxyCompactItemize}

[…]
