### 2025-04-09 8ec7167
add manual link in README
