### 2025-04-09 abeb1b9
add automatic test scripts
