### 2025-04-09 itsoulos/GlobalOptimus#6 @itsoulos
@diehlpk  and @PaulDebus  and @nicoguaro
We have added three scripts to perform automatic test
1. The script ./single_test.sh is used to test an objective problem by applying a series of optimization methods on it. The user can invoke it using the example: ./single_test.sh bf1 and a series of methods will be applied on bf1 test problem.
2. The script ./alll_test.sh applies the same list of optimization methods to all problems of the distribution having a known global minimum.
3. The script ./install_and_test.sh can be used to compile the software and apply a series of optimization methods on a selected test problem. The user can execute as in the example: ./install_and_test.sh bf2

Also, we have updated accordingly the README.md by adding a new subsection named Automatic Testing.

Finally, we have added documentation using comments in every class of the distribution and we have produced a doxygen manual in the subdirectory MANUAL of the distribution. Also, a static link that directs the user to the manual has been added in the beginning of README.md
