### 2025-04-09 #7584 @diehlpk
@PaulDebus and @nicoguaro I think the automated tests are fine. Please have a look and check the box if you agree.
