### 2025-04-12 itsoulos/GlobalOptimus#6 @PaulDebus
yes, that looks much better. i will make a small PR to the README, but i think this is fine. thanks
