### 2025-04-12 itsoulos/GlobalOptimus#1 @PaulDebus
All the issues I opened are now resolved, I think the paper can now be published. Thanks for being so cooperative during the process and integrating our suggestions.

I have some more ideas and thoughts for you, please take them as possible next steps if you continue working on this project:
- You could reduce the visual presence of the GUI application in the README. As far as I can tell, the software works just fine without it and needs to be recompiled to integrate user problems and algorithms, so installing the software is not really the important step
- You could restructure the README a bit to have clearer sections and more focus on the relevant parts - there is now a lot of information in there
- The README contains some explicit formatting that might not be required. Also, you sometimes missed putting an empty line to separate paragraphs. I think this could be simplified to ease readability and maintainability.
- As mentioned in the review issue in the JOSS repo, you could give a simpler, step-by-step example of how to implement a user problem. Which file to put where, what to put in there, how to recompile etc...
- You could link the JOSS paper (once it is complete) in the README
- You could extend a bit the validation of the performance of the optimization algorithms by comparing them to known optima. This can give a lot of really important and helpful information to users to show that optima can be found reliably
