### 2025-04-12 #7584 @PaulDebus
Thanks everyone for working on this, I think the paper has reached a good shape. The points that @nicoguaro raised above are important and relevant, but from my side, this is fine.

I would suggest a last round of language and grammar checking for both the paper and the README, but this is not blocking.
