### 2025-04-16 #7584 @diehlpk
@itsoulos

Please do the following

- [ ] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. Post the release number here.
Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper. Please post the DOI here.
