### 2025-04-16 #7584 @itsoulos
@diehlpk

1) Here is the release link: https://github.com/itsoulos/GlobalOptimus/releases/tag/v1.0.0
2) Here is the zenodo DOI: 10.5281/zenodo.15226967
3) We have checked the license and it is the same as in zenodo

Thank you for your assistance
