### 2025-04-16 #7584 @diehlpk
@itsoulos I just finished reading the paper and found some minor issue

-  https://qt.io -> Pleae consider making that a link
- Above Implemented global optimization methods -> please remove the `width=50%} `
- Remove the `width=50%} ` below the flow chart
- The flow chart could be resized.
- Line 58 remove ??
- Some figures have captions and some not. Maybe make it consistent
