### 2025-04-16 #7584 @diehlpk
@itsoulos The ISO C++ website has some hints to have a nicer C++ in LaTex, see here https://isocpp.org/wiki/faq/misc-environmental-issues#latex-macros

Not a deal breaker, but it will look nicer.
