+++ paper/paper.md
@@ -49,7 +49,7 @@ Finally, Gurobi is an extremely powerful tool for mathematical optimization, but
 
 # Usage of the Software 
 <p style="text-align: justify;">
-The software is compiled using the QT library, available from  [Qt site](https://qt.io) . The researcher should code their objective function and a number of other mandatory functions in the C++ programming language, specifying the dimension of the objective function and the bounds of the function domain. Subsequently, they can select a global optimization method to apply to the problem from a wide range of available methods.</p>
+The software is compiled using the QT library, available from  [Qt main url](https://qt.io) The researcher should code their objective function and a number of other mandatory functions in the C++ programming language, specifying the dimension of the objective function and the bounds of the function domain. Subsequently, they can select a global optimization method to apply to the problem from a wide range of available methods.</p>
 
 In addition, the user can create his own objective function by implementing the methods of category `UserProblem`,
 the methods of which are presented in the following diagram.
