### 2025-04-16 #7584 @itsoulos
@diehlpk  I have added the link and I think that captions are visible now in figures
