### 2025-04-16 #7584 @itsoulos
@diehlpk  also the image has been resized. Is there any command that I can use in the md file to scale it more?
