### 2025-04-16 #7584 @diehlpk
@itsoulos

Figure sizes can be customized by adding an optional second parameter:
`![Caption for example figure.](figure.png){ width=20% }`

More details can be found here https://joss.readthedocs.io/en/latest/example_paper.html
