### 2025-04-16 #7584 @diehlpk
@itsoulos Figure 1 looks blurry. Can you fix that?

The submission is ready and we can accept the paper once the cosmetic issues are fixed.
