+++ paper/paper.md
@@ -103,7 +103,7 @@ Method   |  DESCRIPTION |
 A flowchart of any used optimization process provided below:
 
 
-![The diagram of the optimization process ](usermethod.png){ width=20% }
+![The diagram of the optimization process ](usermethod.png){ width=60% }
 
 An example run for the Rastrigin function and the Genetic algorithm is outlined in \autoref{FIG:rastrigin}.
 
