### 2025-04-17 #7584 @itsoulos
@diehlpk  I have fixed the problem with the blur image and I have also added caption to the first image of the article.
