### 2025-04-17 #7584 @itsoulos
Dear @danielskatz
1) I have merged your pull changes. Thanks
2) The OPTIMUS was the old name of the package where we used shared libraries to load functions and methods, but we decided to create a new project where the packages are available as c++ class that are compiled statically with the rest of package
3) We did not receive any funding for our project.
