### 2025-04-17 #7584 @danielskatz
> 2. The OPTIMUS was the old name of the package where we used shared libraries to load functions and methods, but we decided to create a new project where the packages are available as c++ class that are compiled statically with the rest of package

Does it make sense to leave the title as it is then?  Or would it be better to change the title to match the new name of the package?
