### 2025-04-17 #7584 @itsoulos
@danielskatz  we have abandoned the initial OPTIMUS package, but we have used it as a name in many papers that we have published during the past 3-4 years.  If there is no problem on your part, we would like to leave it with this name.
