### 2025-04-17 #7584 @danielskatz
It's fine with me, though I think it will be confusing to readers, but it's up to you.  I'll create another draft now. Please confirm that it looks good, and if it does to me as well, I will then proceed to accept and publish the work.
